% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{addIDCols}
\alias{addIDCols}
\title{Add ID columns back to \code{downscale} output.}
\usage{
addIDCols(IDCols, results)
}
\arguments{
\item{IDCols}{character. ID columns to add, or NULL if none}

\item{results}{\code{data.table} or \code{SpatVector} output from \code{downscale}}
}
\value{
\code{results} with IDCols
}
\description{
Add ID columns back to \code{downscale} output.
}
\details{
adds back the original ID column and removes temporary
one
}
