% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{ahccd_pstations}
\alias{ahccd_pstations}
\title{Adjusted Precipitation for Canada}
\format{
a \code{PackedSpatVector}
}
\source{
\url{https://www.canada.ca/en/environment-climate-change/services/climate-change/science-research-data/climate-trends-variability/adjusted-homogenized-canadian-data/precipitation.html}
}
\usage{
ahccd_pstations
}
\description{
A points shapefile of the Adjusted Precipitation for Canada
(APC2) dataset for over 462 locations.
}
\details{
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{ahccd_pstations <- terra::unwrap(get(data(ahccd_pstations)))}
}
\keyword{datasets}
