% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{append_calc}
\alias{append_calc}
\title{Add extra climate variables to a data.table}
\usage{
append_calc(dt, extra_variables, grouping)
}
\arguments{
\item{dt}{A data.table with TminXX, TmaxXX, PPTXX for XX in 01 to 12.}

\item{extra_variables}{A character vector of extra variables to compute.}

\item{grouping}{A character vector of variables grouping. Can be \code{m} (monthly), \code{s} (seasonal) or \code{a} (annual).}
}
\description{
Add extra climate variables to a data.table
}
