% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{append_extra}
\alias{append_extra}
\title{Add extra climate variables to a data.table}
\usage{
append_extra(dt, extra)
}
\arguments{
\item{dt}{A data.table with TminXX, TmaxXX, PPTXX for XX in 01 to 12.}

\item{extra}{A character vector of extra variables to compute.}
}
\description{
Add extra climate variables to a data.table
}
