% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{baseline}
\alias{baseline}
\title{Create normal baseline for \code{downscale}}
\usage{
baseline(normal = list_normal()[1], dem = list_dem()[1])
}
\arguments{
\item{normal}{A character. Label of the normal baseline to use. Can be obtained from
\code{list_normal()}. Default to \code{list_normal()[1]}.}

\item{dem}{A character. Label of the digital elevation model to use. Must have the same
extent as the normal baseline. Can be obtained from \code{list_dem()}. Default to \code{list_dem()[1]}.}
}
\value{
A normal baseline to use with \code{downscale}. A \code{SpatRaster} with a \code{dem} attribute.
}
\description{
Create normal baseline for \code{downscale}
}
