% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{BECcols}
\alias{BECcols}
\title{Colour sscheme for Biogeoclimatic Ecological Classification Zones}
\format{
a \code{data.table}
}
\source{
colour scheme commonly used by the BC provincial government.
}
\usage{
BECcols
}
\description{
A \code{data.table} with RGB and HEX colours for biogeoclimatic zones in BC.
}
\keyword{datasets}
