% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{BECz_vancouver}
\alias{BECz_vancouver}
\title{Example Biogeoclimatic Ecological Classification Zones}
\format{
a \code{SpatVector}
}
\source{
obtained from \code{bcdata} package, using \code{bcdc_get_data("bec-map-generalized-1-20k-")}
}
\usage{
BECz_vancouver
}
\description{
A shapefile with biogeoclimatic zones in the coast of Vancouver, BC, Canada.
}
\keyword{datasets}
