% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{BECz_vancouver_ras}
\alias{BECz_vancouver_ras}
\title{Example Biogeoclimatic Ecological Classification Zones in raster format}
\format{
a \code{PackedSpatRaster}
}
\source{
derived from \code{BECz_vancouver}
}
\usage{
BECz_vancouver_ras
}
\description{
A categorical raster of biogeoclimatic zones in the coast of Vancouver, BC, Canada.
Derived from \code{BECz_vancouver} using \code{dem_vancouver} as the template for extent,
resolution and projection information.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{BECz_vancouver_ras <- terra::unwrap(get(data(BECz_vancouver_ras)))}
}
\keyword{datasets}
