% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear the package's local cache path}
\usage{
cache_clear(what = c("gcm", "normal", "historic"))
}
\arguments{
\item{what}{character. Which data folders should be cleared?
Accepts "normal", "gcm" or both.}
}
\description{
Attempts to delete all folder/files in \code{cache_path()}.
}
\details{
It may fail if R has no permission to delete files/folders
in the \code{cache_path()} directory
}
