% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_path}
\alias{cache_path}
\title{Return package local cache path}
\usage{
cache_path()
}
\value{
character. The full path of the package local cache.
}
\description{
Return package local cache path
}
\details{
By default, it uses \code{\link[tools:userdir]{tools::R_user_dir()}}. The cache location can be
set using the \code{climr.cache.path} option with \code{options("climr.cache.path" = "your_path")}.
}
