% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_CMD}
\alias{calc_CMD}
\title{Calculate climatic moisture deficit (CMD)}
\usage{
calc_CMD(Eref, PPT)
}
\arguments{
\item{Eref}{Reference evaporation}

\item{PPT}{numeric. Precipitation in mm.}
}
\value{
numeric. Climatic moisture deficit
}
\description{
Calculate climatic moisture deficit (CMD)
}
