% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DD.R
\name{calc_DD_above_18}
\alias{calc_DD_above_18}
\title{Calculate Degree-Day Above 18 (DD>18)}
\usage{
calc_DD_above_18(m, tm, region)
}
\arguments{
\item{m}{integer. Month of the year.}

\item{tm}{numeric. Monthly mean temperature for the \code{m} month.}

\item{region}{character. One of either "All", "South west", "The rest".}
}
\value{
numeric. Degree-Days Above 18
}
\description{
Calculate Degree-Day Above 18 (DD>18)
}
\examples{
\dontrun{
climr:::calc_DD_above_18(2, -14, "All")
}
}
