% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DD.R
\name{calc_DD_above_5}
\alias{calc_DD_above_5}
\title{Calculate Degree-Day Above 5 (DD>5)}
\usage{
calc_DD_above_5(m, tm, region)
}
\arguments{
\item{m}{month of the year}

\item{tm}{monthly mean temperature for the \code{m} month}

\item{region}{one of either "All", "West", "East"}
}
\value{
Degree-Day Above 5
}
\description{
Calculate Degree-Day Above 5 (DD>5)
}
\examples{
calc_DD_above_5(2, -14, "All")
}
