% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DD.R
\name{calc_DD_below_18}
\alias{calc_DD_below_18}
\title{Calculate Degree-Day Below 18 (DD<18)}
\usage{
calc_DD_below_18(m, tm)
}
\arguments{
\item{m}{month of the year}

\item{tm}{monthly mean temperature for the \code{m} month}
}
\value{
Degree-Days Below 18
}
\description{
Calculate Degree-Day Below 18 (DD<18)
}
\examples{
\dontrun{
climr:::calc_DD_below_18(2, -14)
}
}
