% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_FFP.R
\name{calc_eFFP}
\alias{calc_eFFP}
\title{Calculate eFFP}
\usage{
calc_eFFP(NFFD, t_min_list)
}
\arguments{
\item{NFFD}{Number of Frost-Free Days}

\item{t_min_list}{named list of monthly minimum temperature for each month}
}
\value{
Day of the year on which the Frost-Free Period ends
}
\description{
bFFP : Day of the year on which the Frost-Free Period ends
}
\examples{
\dontrun{
t_min_list <- list("1" = -35, "2" = -32, "3" = -25, "4" = -10, 
"5" = -5, "6" = 3,  "7" = 15, "8" = 17, "9" = 10, "10" = -5,
"11" = -20, "12" = -30)

climr:::calc_eFFP(NFFD = 10, t_min_list = t_min_list)
}
}
