% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_Eref}
\alias{calc_Eref}
\title{Calculate Extreme Minimum Temperature (EMT)}
\usage{
calc_Eref(m, tmmin, tmmax, latitude)
}
\arguments{
\item{m}{integer. Month of the year.}

\item{tmmin}{numeric. Monthly mean minimum air temperature.}

\item{tmmax}{numeric. Monthly mean maximum air temperature.}

\item{latitude}{numeric. Point latitudes.}
}
\value{
numeric. Reference evaporation (Eref)
}
\description{
Calculate Extreme Minimum Temperature (EMT)
}
