% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EMT_EXT.R
\name{calc_EXT}
\alias{calc_EXT}
\title{Calculate Extreme Maximum Temperature}
\usage{
calc_EXT(t_max_list, td)
}
\arguments{
\item{t_max_list}{list of monthly maximum temperature for each month in order}

\item{td}{difference between the mean warmest monthly temperature and the mean coldest monthly temperature}
}
\value{
Extreme Maximum Temperature
}
\description{
Calculate Extreme Maximum Temperature
}
