% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_FFP.R
\name{calc_FFP}
\alias{calc_FFP}
\title{Calculate Frost-Free Period}
\usage{
calc_FFP(bFFP, eFFP)
}
\arguments{
\item{bFFP}{numeric. Day of the year on which the Frost-Free Period begins}

\item{eFFP}{numeric. Day of the year on which the Frost-Free Period ends}
}
\value{
numeric. Frost-Free Period in days
}
\description{
Calculate Frost-Free Period
}
\examples{
\dontrun{
climr:::calc_FFP(bFFP = 214.5964, eFFP = 265.4581)
}
}
