% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NFFD.R
\name{calc_NFFD}
\alias{calc_NFFD}
\title{Calculate Number of Frost Free Days (NFFD)}
\usage{
calc_NFFD(m, tm)
}
\arguments{
\item{m}{month of the year}

\item{tm}{min temperature for that month}
}
\value{
Number of Frost Free Days
}
\description{
Calculate Number of Frost Free Days (NFFD)
}
\examples{
\dontrun{
climr:::calc_NFFD(3, 2.05)
}
}
