% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PAS.R
\name{calc_PAS}
\alias{calc_PAS}
\title{Calculate Precipitation As Snow (PAS)}
\usage{
calc_PAS(m, tmin, PPT)
}
\arguments{
\item{m}{integer. Month of the year.}

\item{tmin}{numeric. Minimum temperature for the \code{m} month.}

\item{PPT}{numeric. Precipitation in mm.}
}
\value{
numeric. Precipitation As Snow
}
\description{
Calculate Precipitation As Snow (PAS)
}
\examples{
\dontrun{
climr:::calc_PAS(4, 2, 600)
}
}
