% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RH.R
\name{calc_PET}
\alias{calc_PET}
\title{Calculate Potential Evapotranspiration}
\usage{
calc_PET(tave, tmin, tmax, alt)
}
\arguments{
\item{tave}{numeric. Monthly average minimum air temperature.}

\item{tmin}{monthly mean minimum air temperature}

\item{tmax}{monthly mean maximum air temperature}

\item{alt}{numeric. Altitude in m.}
}
\description{
Based on simplified Penman - Monteith method from Hogg (1997)
}
\references{
Hogg, E.H. (1997). Temporal scaling of moisture and the forest-grassland boundary in western Canada. Agricultural and Forest Meteorology, Research on Forest Environmental Influences in a Changing World, 84, 115–122.
}
