% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RH.R
\name{calc_RH}
\alias{calc_RH}
\title{Calculate Relative Humidity (RH)}
\usage{
calc_RH(tmmin, tmmax)
}
\arguments{
\item{tmmin}{numeric. Monthly mean minimum air temperature.}

\item{tmmax}{numeric. Monthly mean maximum air temperature.}
}
\value{
numeric. Relative Humidity
}
\description{
Calculate Relative Humidity (RH)
}
\examples{
\dontrun{
climr:::calc_RH(tmmin = 10, tmmax = 40)
}
}
