% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_S0_I}
\alias{calc_S0_I}
\title{PROGRAM I From Hargreaves 1985}
\usage{
calc_S0_I(d, tmean, latitude)
}
\arguments{
\item{d}{julian day of the year (January 1 = 1, Decembre 31 = 365)}

\item{tmean}{mean temperature for that month}

\item{latitude}{point latitudes}
}
\value{
Extraterrestrial radiation estimation in mm/day
}
\description{
PROGRAM I From Hargreaves 1985
}
