% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_S0_II}
\alias{calc_S0_II}
\title{PROGRAM II From Hargreaves 1985}
\usage{
calc_S0_II(m, tmean, latitude)
}
\arguments{
\item{m}{month of the year}

\item{tmean}{mean temperature for that month}

\item{latitude}{point latitudes}
}
\value{
Extraterrestrial radiation estimation in mm/day
}
\description{
PROGRAM II From Hargreaves 1985
}
