% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RH.R
\name{calc_SVP}
\alias{calc_SVP}
\title{Calculate Saturated Vapour Pressure at a temperature t}
\usage{
calc_SVP(t)
}
\arguments{
\item{t}{numeric. Air temperature in degree Celsius.}
}
\description{
Calculate Saturated Vapour Pressure at a temperature t
}
