% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DD.R, R/calc_EMT_EXT.R, R/calc_Eref_CMD.R,
%   R/calc_FFP.R, R/calc_NFFD.R, R/calc_PAS.R, R/calc_RH.R
\name{calc_DD_below_0}
\alias{calc_DD_below_0}
\alias{calc_DD_above_5}
\alias{calc_DD_below_18}
\alias{calc_DD_above_18}
\alias{calc_EMT}
\alias{calc_EXT}
\alias{calc_MCMT}
\alias{calc_MWMT}
\alias{calc_TD}
\alias{calc_Eref}
\alias{calc_CMD}
\alias{calc_bFFP}
\alias{calc_eFFP}
\alias{calc_FFP}
\alias{calc_NFFD}
\alias{calc_PAS}
\alias{calc_RH}
\alias{calc_PET}
\alias{calc_SVP}
\title{Climate variables calculation}
\usage{
calc_DD_below_0(m, tm)

calc_DD_above_5(m, tm, region)

calc_DD_below_18(m, tm)

calc_DD_above_18(m, tm, region)

calc_EMT(t_min_list, mcmt, td)

calc_EXT(t_max_list, mwmt, td)

calc_MCMT(t_min_list)

calc_MWMT(t_max_list)

calc_TD(mcmt, mwmt)

calc_Eref(m, tmmin, tmmax, latitude)

calc_CMD(Eref, PPT)

calc_bFFP(td, NFFD, t_min_list)

calc_eFFP(NFFD, t_min_list)

calc_FFP(bFFP, eFFP)

calc_NFFD(m, tmin)

calc_PAS(m, tave, PPT)

calc_RH(tmmin, tmmax)

calc_PET(tave, tmmin, tmmax, alt)

calc_SVP(t)
}
\arguments{
\item{m}{integer. Month of the year.}

\item{tm}{numeric. Monthly mean temperature for the \code{m} month.}

\item{region}{character. One of either "All", "West", "East".}

\item{t_min_list}{\code{list} of monthly minimum temperature for each month in order.}

\item{mcmt}{Mean Temperature of the Coldest Month}

\item{td}{numeric. Difference between the mean warmest monthly temperature and
the mean coldest monthly temperature.}

\item{t_max_list}{\code{list} of monthly maximum temperature for each month in order}

\item{mwmt}{Mean Temperature of the Warmest Month}

\item{tmmin}{numeric. Monthly mean minimum air temperature.}

\item{tmmax}{numeric. Monthly mean maximum air temperature.}

\item{latitude}{numeric. Point latitudes.}

\item{Eref}{Reference evaporation}

\item{PPT}{numeric. Precipitation in mm.}

\item{NFFD}{numeric. Number of Frost-Free Days.}

\item{bFFP}{numeric. Day of the year on which the Frost-Free Period begins}

\item{eFFP}{numeric. Day of the year on which the Frost-Free Period ends}

\item{tmin}{numeric. Minimum temperature for the \code{m} month.}

\item{tave}{numeric. Monthly average minimum air temperature.}

\item{alt}{numeric. Altitude in m.}

\item{t}{numeric. Air temperature in degree Celsius.}
}
\value{
numeric. Reference evaporation (Eref)

numeric. Climatic moisture deficit

numeric. Julian day on which the Frost-Free Period begins

numeric. Julian day on which the Frost-Free Period ends

numeric. Frost-Free Period in days

numeric. Number of Frost Free Days

numeric. Precipitation As Snow

numeric. Relative Humidity
}
\description{
bFFP : Day of the year on which the Frost-Free Period begins

eFFP : Day of the year on which the Frost-Free Period ends

Based on simplified Penman - Monteith method from Hogg (1997)
}
\examples{
\dontrun{
climr:::calc_DD_below_0(2, -14)
}
\dontrun{
climr:::calc_DD_above_5(2, -14, "All")
}
\dontrun{
climr:::calc_DD_below_18(2, -14)
}
\dontrun{
climr:::calc_DD_above_18(2, -14, "All")
}
\dontrun{
t_min_list <- list(
  "1" = -35, "2" = -32, "3" = -25, "4" = -10,
  "5" = -5, "6" = 3, "7" = 15, "8" = 17, "9" = 10, "10" = -5,
  "11" = -20, "12" = -30
)

climr::calc_bFFP(td = 30, NFFD = 10, t_min_list = t_min_list)
}
\dontrun{
t_min_list <- list(
  "1" = -35, "2" = -32, "3" = -25, "4" = -10,
  "5" = -5, "6" = 3, "7" = 15, "8" = 17, "9" = 10, "10" = -5,
  "11" = -20, "12" = -30
)

climr::calc_eFFP(NFFD = 10, t_min_list = t_min_list)
}
\dontrun{
climr::calc_FFP(bFFP = 214.5964, eFFP = 265.4581)
}
\dontrun{
climr::calc_NFFD(3, 2.05)
}
\dontrun{
climr::calc_PAS(4, 2, 600)
}
\dontrun{
climr::calc_RH(tmmin = 10, tmmax = 40)
}
}
\references{
Hogg, E.H. (1997). Temporal scaling of moisture and the forest-grassland boundary in western Canada. Agricultural and Forest Meteorology, Research on Forest Environmental Influences in a Changing World, 84, 115–122.
}
