% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-package.R
\docType{package}
\name{climr-package}
\alias{climr}
\alias{climr-package}
\title{climr: Downscaling of global climate data}
\description{
This package replicates the basic approach of ClimateNA (climatena.ca), but with (1) improved speed, (2) multiple runs of each GCM, (3) cloud delivery, and (4) programmatic access via R. A simple downscaling method is used to bias-correct low-resolution (~100km) global climate model anomalies to high-resolution (2km) maps of historical climate, with further elevation adjustment to user-specified elevation grids/points based on empirical lapse rates (local relationship of climate to elevation) of the 2km climate maps. Elevation-adjusted monthly values of basic climate elements (temperature and precipitation) are then used to estimate derived variables (e.g., degree-days) based on published equations and parameters from Wang et al. 2016.
}
\author{
\strong{Maintainer}: Colin Mahony \email{Colin.Mahony@gov.bc.ca} (\href{https://orcid.org/0000-0002-6111-5675}{ORCID})

Authors:
\itemize{
  \item Kiri Daust \email{kiri.daust@gov.bc.ca}
  \item Bruno Tremblay \email{bruno@boostao.ca} (\href{https://orcid.org/0000-0002-2945-356X}{ORCID})
  \item Ceres Barros \email{ceres.barros@gov.bc.ca} (\href{https://orcid.org/0000-0003-4036-977X}{ORCID})
  \item Francois Bornais \email{francois@boostao.ca}
}

Other contributors:
\itemize{
  \item  Province of British Columbia [copyright holder, funder]
}

}
\keyword{internal}
