% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{climr_downscale}
\alias{climr_downscale}
\title{Downscale and Calculate climate variables for points of interest}
\usage{
climr_downscale(
  xyz,
  which_normal = c("auto", "BC", "NorAm"),
  historic_period = NULL,
  historic_ts = NULL,
  gcm_models = NULL,
  ssp = c("ssp126", "ssp245", "ssp370", "ssp585"),
  gcm_period = c("2001_2020", "2021_2040", "2041_2060", "2061_2080", "2081_2100"),
  gcm_ts_years = 2020:2030,
  max_run = 0L,
  return_normal = TRUE,
  vars = sort(sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))),
  cache = TRUE
)
}
\arguments{
\item{xyz}{three or four column data.frame: long, lat, elev, (id)}

\item{which_normal}{Select which normal layer to use. Default is "auto", which selected the highest resolution normal for each point}

\item{historic_period}{Which historic period. Default \code{NULL}}
}
\description{
Downscale and Calculate climate variables for points of interest
}
