% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr.R
\docType{data}
\name{cmip6_pr}
\alias{cmip6_pr}
\title{CMIP6 21st Century Precipitation Projection Data}
\format{A data frame with 7268 rows and 6 columns.
  \itemize{
    \item{Column 1, \code{index}: }{percentage anomaly of the 2020-2049 mean
      (if \code{period = 1} in column 6) or the 2070-2099 mean
      (if \code{period = 2} in column 6) relative to the 1970-1999 mean.
      That is, the latter subtracted from the former divided by former,
      expressed as a percentage.}
    \item{Column 2, \code{GCM}: }{Abbreviated name of General Circulation
      Model.  There are 11 different GCMs.}
    \item{Column 3, \code{SSP}: }{Shared Socio-Economic Pathway.
      One of ssp119, ssp126, ssp245, ssp370, ssp434, ssp460, ssp534-over,
      ssp585.}
    \item{Column 4, \code{run}: }{Indicator of the simulation run number.}
    \item{Column 5, \code{region}: }{Region of the world: 0 for global,
      1-22 for regions.}
    \item{Column 6, \code{period}: }{Time horizon: 1 for mid 21st century
      (2020-2049); 2 for late 21st century (2070-2099).}
 }}
\source{
The raw data from which the indices are calculated were downloaded
  via the World Climate Research Programme (WCRP) CMIP6 homepage
  \url{https://esgf-node.llnl.gov/projects/cmip6/}.
}
\usage{
cmip6_pr
}
\description{
Indices of percentage precipitation change from late 20th century (1970-1999)
to mid 21st century (2020-2049) and to late 21st century (2070-2099) based
on data produced by the Sixth Coupled Model Intercomparison Project (CMIP6).
The climate variable is precipitation flux (pr).
}
\details{
The data frame \code{cmip6_pr} data frame has 7268 rows and 6 columns.
Each row relates to a climate projection run from one of 11 different
General Circulation Models (GCMs) under a particular Shared Socio-Economic
Pathway (SSP) for a particular region of the world (and one of two time
horizons.

See Northrop and Chandler (2014) for analyses of a similar older dataset
(CMIP3).
}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying
  Sources of Uncertainty in Projections of Future Climate.
  \emph{Journal of Climate}, \strong{27}, 8793-8808.
  \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\keyword{datasets}
