% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries.R
\name{colSelect}
\alias{colSelect}
\title{function for specifying the color}
\usage{
colSelect(scenario, gcm, pal.scenario, scenarios, pal, pal.gcms)
}
\arguments{
\item{scenario}{TODO}

\item{gcm}{TODO}

\item{pal.scenario}{TODO}

\item{scenarios}{TODO}

\item{pal}{character. color palette. Options are "scenario", for use when comparing scenarios,
and "gcms", for use when comparing GCMs.}

\item{pal.gcms}{TODO}
}
\description{
function for specifying the color
}
\keyword{internal}
