% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{content_get}
\alias{content_get}
\title{Get file content from Github}
\usage{
content_get(org, repo, file, ref = NULL)
}
\arguments{
\item{org}{Organization name}

\item{repo}{Repo name}

\item{file}{File path relative to repo root without starting /}

\item{ref}{The name of the commit/branch/tag. Default: the repository’s default branch}
}
\value{
File content as character
}
\description{
Get file content from Github
}
