% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create a Boxplot from Climate Data}
\usage{
create_boxplot(dt, var = c("Tmax", "Tmin", "PPT", "Tave"))
}
\arguments{
\item{dt}{A data.table containing the climate data with columns for monthly values
and optionally a RUN column.}

\item{var}{Character string specifying the variable to plot. One of:
"Tmax" (maximum temperature), "Tmin" (minimum temperature),
"PPT" (precipitation), or "Tave" (average temperature).}
}
\value{
A ggplot object containing a boxplot of monthly values by period.
}
\description{
Creates a boxplot visualizing monthly climate data (temperature or precipitation)
across different periods using ggplot2.
}
\details{
The function processes climate data by calculating means by period using ensemble means
if present and then create the boxplot
}
\examples{

# Sample data
dt <- data.frame(
  PERIOD = rep(c("2000-2010", "2011-2020"), each = 2),
  Tmax_01 = c(10, 12, 11, 13),  
  Tmax_02 = c(15, 16, 14, 17),  
  Tmax_03 = c(18, 19, 17, 20),  
  Tmax_04 = c(22, 23, 21, 24),  
  Tmax_05 = c(25, 26, 24, 27),  
  Tmax_06 = c(28, 29, 27, 30),  
  Tmax_07 = c(30, 31, 29, 32), 
  Tmax_08 = c(29, 30, 28, 31),  
  Tmax_09 = c(26, 27, 25, 28),  
  Tmax_10 = c(22, 23, 21, 24), 
  Tmax_11 = c(17, 18, 16, 19), 
  Tmax_12 = c(12, 13, 11, 14)  
  )
  
# Create boxplot
create_boxplot(dt, "Tmax")


}
