% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_climate_diagram.R
\name{create_climate_diagram}
\alias{create_climate_diagram}
\title{Create Climate Diagram Plot}
\usage{
create_climate_diagram(temp, precip, elev = NULL)
}
\arguments{
\item{temp}{vector of temperatures (in Celsius) for each month}

\item{precip}{vector of precipitation (in mm) for each month}

\item{elev}{elevation (in m). Optional, if NULL no elevation will be shown in output.}
}
\description{
Create Climate Diagram Plot
}
\examples{

temperature <- c(14, 14, 17, 20, 24, 28, 31, 31, 27, 23, 18, 15)
precipitations <- c(93.1, 62.1, 58.8, 46.6, 45.9, 23.1, 11.2, 18.9, 73.8, 68, 155.3, 76.3)
create_climate_diagram(temp = temperature, precip = precipitations,  elev = 25)

create_climate_diagram(temp = c(-3, 14, 17, 20, 24, 28, 31, 31, 27, 23, 18, -5), 
   precip = c(93.1, 62.1, 58.8, 46.6, 45.9, 23.1, 11.2, 18.9, 73.8, 68, 155.3, 76.3), 
   elev = 25 )
   
\dontrun{  
 in_xyz <- data.frame(lon = -127.7052, lat = 55.3557, elev = 291, id = 1)
 diagram_info <- create_climate_diagram_input(in_xyz)
 create_climate_diagram(temp = diagram_info[["Tave"]], precip = diagram_info[["PPT"]], elev = diagram_info[["elev"]])
}

}
