% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_diagram.R
\name{create_climate_diagram_input}
\alias{create_climate_diagram_input}
\title{Create input for Climate Diagram}
\usage{
create_climate_diagram_input(
  xyz,
  gcms = list_gcms()[c(1, 4, 5, 6, 7, 10, 11, 12)],
  ssps = list_ssps()[2],
  obs_periods = list_obs_periods()[1],
  gcm_periods = list_gcm_periods(),
  max_run = 10L,
  cache = TRUE,
  use_downscale_db = FALSE,
  ...
)
}
\arguments{
\item{xyz}{a \code{terra::SpatRaster} or a \code{data.frame} with the following columns
"long", "lat", "elev", and a unique "id". Any extra columns will be ignored and not output.}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{obs_periods}{character. Which historical period for observational climate
data, averaged over this period. Options are \code{\link[=list_obs_periods]{list_obs_periods()}}.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{use_downscale_db}{Should the function \code{downscale_db} be used instead of \code{downscale}}

\item{...}{other arguments passed to \code{\link[=downscale_core]{downscale_core()}}. Namely: \code{return_refperiod},
\code{vars}, \code{out_spatial} and \code{plot}}
}
\description{
This function will downscale and return a list of the montly Tmin, Tmax, Tave, PPT and the elevation.
}
\examples{
\dontrun{
in_xyz <- data.frame(lon = -127.7052, lat = 55.3557, elev = 291, id = 1)
create_climate_diagram_input(in_xyz)
}
}
