% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lists.R
\name{list_gcms}
\alias{list_gcms}
\alias{list_ssps}
\alias{list_gcm_periods}
\alias{list_runs_ssp}
\alias{list_runs_historic}
\alias{list_refmaps}
\alias{list_obs_periods}
\alias{list_vars}
\alias{list_obs_years}
\alias{list_gcm_ssp_years}
\alias{list_gcm_hist_years}
\title{List available runs, global circulation models, periods and climate scenarios}
\usage{
list_gcms()

list_ssps()

list_gcm_periods()

list_runs_ssp(gcm, ssp)

list_runs_historic(gcm)

list_refmaps()

list_obs_periods()

list_vars(set = c("All", "Monthly", "Seasonal", "Annual"), only_extra = FALSE)

list_obs_years()

list_gcm_ssp_years()

list_gcm_hist_years()
}
\arguments{
\item{gcm}{Name of GCM}

\item{ssp}{Name of scenario}

\item{set}{character. One of All, Monthly, Seasonal, Annual, or any combination thereof. Defaults to "All".}

\item{only_extra}{logical. Should Tmin, Tmax and PPT be excluded? Defaults to FALSE.}
}
\value{
a character vector.
}
\description{
\code{list_gcms} lists available global circulation models.

\code{list_ssps} lists available shared socioeconomic pathways.

\code{list_gcm_periods} lists available periods.

lists available runs for a given GCM/ssp.

lists available runs for a given historic GCM.

\code{list_refmaps} lists available reference maps of gridded climate normals

\code{list_obs_periods} lists available observational periods

\code{list_vars} lists climate variables

\code{list_obs_years} lists available years for obs time series

\code{list_gcm_ssp_years} lists available years for future projections' time series

\code{list_gcm_hist_years} lists available years for obs projections' time series
}
\details{
Currently available reference maps of gridded climate normals (\code{list_refmaps()}) are:
\itemize{
\item "refmap_climatena" for Climate NA derived normals
\item "refmap_prism" for British Columbia PRISM climatologies derived normals
\item "refmap_climr" for a composite of BC PRISM, adjusted US PRISM and
DAYMET (Alberta and Saskatchewan), covering western Canada and western
US.
}
}
