% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{data_con}
\alias{data_con}
\alias{connections}
\alias{connections_clear}
\title{List connections in cache}
\usage{
data_con(profile = c(.globals$last_profile, "climr-db-user", "local"))

connections()

connections_clear()
}
\arguments{
\item{profile}{Either \code{climr-db-user} or \code{local}.}
}
\description{
List connections in cache

List connections in cache

Clear connections in cache
}
