% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{data_connect}
\alias{data_connect}
\title{Connect to PostGIS database}
\usage{
data_connect(local = FALSE)
}
\arguments{
\item{local}{A logical. Use a local database. Default \code{FALSE}.}
}
\value{
pool object of database connection
}
\description{
Connect to PostGIS database
}
