% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_delete}
\alias{data_delete}
\title{Delete package local cache}
\usage{
data_delete(ask = interactive())
}
\arguments{
\item{ask}{A boolean. Ask before deleting files. Default to \code{interactive()}.}
}
\description{
Delete package local cache
}
