% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_download}
\alias{data_download}
\title{Download file to data path}
\usage{
data_download(url, path, uid, quiet = !interactive())
}
\arguments{
\item{url}{A character. Remote file location}

\item{path}{A character. Path where to save file relative to data path}

\item{uid}{A character. A unique identifier used to determine if file has been
updated or not.}

\item{quiet}{A logical. If \code{TRUE}, suppress status messages (if any), and the progress bar.}
}
\description{
Download file to data path
}
