% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_get}
\alias{data_get}
\title{Process a list of files.}
\usage{
data_get(files, quiet = !interactive())
}
\arguments{
\item{files}{A list of files.}

\item{quiet}{A logical. If \code{TRUE}, suppress status messages (if any), and the progress bar.}
}
\description{
Process a list of files.
}
\details{
Each element must have a \code{url}, a \code{path} relative
to data path and a unique identifier \code{uid}.
}
