% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_path}
\alias{data_path}
\title{Return current package data path}
\usage{
data_path()
}
\description{
Return current package data path
}
\details{
If cache is used, return the cache path. Otherwise, it returns
a temporary folder that will be used for this session only.
}
