% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{dbGetRaster}
\alias{dbGetRaster}
\title{Download raster with bounding box from PostGIS (boostao version)}
\usage{
dbGetRaster(conn, name, tile, rast = "rast", bands = 37:73, boundary)
}
\arguments{
\item{conn}{a DBI or RPostgres connection object.}

\item{name}{character. Table name in database.}

\item{tile}{Logical. Retrieve data in tiles to avoid overloading the database?}

\item{rast}{character. Name of column which stores raster data.
Defaults to "rast"}

\item{bands}{numeric. Which raster bands to return. Default 37:73.}

\item{boundary}{numeric. Vector of length 4 (ymax,ymin,xmax,xmin). Defaults to \code{NULL}.}
}
\value{
A \code{SpatRaster}
}
\description{
Download raster with bounding box from PostGIS (boostao version)
}
