% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{dem_vancouver}
\alias{dem_vancouver}
\title{Example DEM}
\format{
a \code{SpatRaster}
}
\source{
todo add/describe source
}
\usage{
dem_vancouver
}
\description{
Digital elevation model of North Vancouver, BC, Canada.
}
\keyword{datasets}
