% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{dem_vancouver_lowres}
\alias{dem_vancouver_lowres}
\title{Example low resolution DEM}
\format{
a \code{PackedSpatRaster}
}
\source{
todo add/describe source
}
\usage{
dem_vancouver_lowres
}
\description{
A lower resolution version of \code{dem_vancouver},
re-projected at 0.005 degrees.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{dem_vancouver_lowres <- terra::unwrap(get(data(dem_vancouver_lowres)))}
}
\keyword{datasets}
