% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{.checkXYZ}
\alias{.checkXYZ}
\title{Check xyz conforms to standards}
\usage{
.checkXYZ(xyz, expectedCols = c("lon", "lat", "elev", "id"))
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{expectedCols}{character. Columns to look for in \code{xyz}.}
}
\description{
Check xyz conforms to standards
}
