% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale target rasters to points of interest}
\usage{
downscale(
  xyz,
  normal,
  gcm = NULL,
  variables,
  grouping = c("m", "s", "a"),
  use_cache = TRUE
)
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{normal}{Reference historical climate variables rasters.}

\item{gcm}{Future climate variables rasters. Default to NULL.}

\item{variables}{A character vector of variables to compute.}

\item{grouping}{A character vector of variables grouping. Can be \code{m} (monthly), \code{s} (seasonal) or \code{a} (annual).}

\item{use_cache}{A boolean. For lapse rate calculation, if \code{TRUE} and available,
uses cached computation. If \code{FALSE}, recompute. Default to \code{TRUE}.}
}
\description{
Downscale target rasters to points of interest
}
