% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{climr_downscale}
\alias{climr_downscale}
\alias{downscale}
\title{Climate variable downscaling}
\usage{
climr_downscale(
  xyz,
  which_normal = c("auto", list_normal()),
  historic_period = NULL,
  historic_ts = NULL,
  gcm_models = NULL,
  ssp = c("ssp126", "ssp245", "ssp370", "ssp585"),
  gcm_period = NULL,
  gcm_ts_years = NULL,
  gcm_hist_years = NULL,
  max_run = 0L,
  return_normal = TRUE,
  vars = sort(sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))),
  cache = TRUE,
  out_spatial = FALSE,
  plot = NULL
)

downscale(
  xyz,
  normal,
  gcm = NULL,
  historic = NULL,
  gcm_ts = NULL,
  gcm_hist = NULL,
  historic_ts = NULL,
  return_normal = FALSE,
  vars = sort(sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))),
  ppt_lr = FALSE,
  nthread = 1L,
  out_spatial = FALSE,
  plot = NULL
)
}
\arguments{
\item{xyz}{A 3-column \code{matrix} or \code{data.frame} (x, y, z) or (lon, lat, elev).}

\item{which_normal}{character. Which normal layer to use.
Default is "auto", which selects the highest resolution normal for each point.
Other options are one of \code{\link[=list_normal]{list_normal()}}}

\item{historic_period}{character. Which historic period. Default \code{NULL}}

\item{historic_ts}{\code{list} of \code{SpatRasters}. TODO}

\item{gcm_models}{character. Vector of GCM names. Options are \code{\link[=list_gcm]{list_gcm()}}. Used for gcm periods, gcm timeseries, and historic timeseries. Default \code{NULL}}

\item{ssp}{character. Vector of labels of the shared socioeconomic pathways to use.
See available options with \code{\link[=list_ssp]{list_ssp()}}. Defaults to all SSPs available.}

\item{gcm_period}{character. Requested future periods. Options are \code{\link[=list_gcm_period]{list_gcm_period()}}}

\item{gcm_ts_years}{character. Requested future timeseries years. Must be in \code{2015:2100}}

\item{gcm_hist_years}{character. Requested historic modelled years. Must be in \code{1851:2010}}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{return_normal}{logical. Return downscaled normal period (1961-1990)? Default \code{TRUE}.}

\item{vars}{character. A vector of climate variables to compute. Supported variables
can be obtained with \code{\link[=list_variables]{list_variables()}}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{out_spatial}{logical. Should a SpatVector be returned instead of a
\code{data.frame}.}

\item{plot}{character. If \code{out_spatial} is TRUE, the name of a variable to plot.
If the variable exists in \code{normal}, then its normal values will also be plotted.
Otherwise, normal January total precipitation (PPT01) values will be plotted.
Defaults to no plotting (NULL).}

\item{normal}{\code{SpatRaster}. Reference normal baseline input from \code{normal_input}.}

\item{gcm}{\code{list} of \code{SpatRasters}. Global Circulation Models outputs from \code{\link[=gcm_input]{gcm_input()}}. Default to \code{NULL}.}

\item{historic}{\code{list} of \code{SpatRasters}. Historic time period outputs from \code{\link[=historic_input]{historic_input()}}. Default to \code{NULL}.}

\item{gcm_ts}{\code{list} of \code{SpatRasters}. TODO}

\item{gcm_hist}{\code{list} of \code{SpatRasters}. TODO}

\item{ppt_lr}{logical. Apply lapse rate adjustment to precipitations. Default to FALSE.}

\item{nthread}{integer. Number of parallel threads to use to do computations. Default to 1L.}
}
\value{
\code{data.frame} of downscaled climate variables for each location.
Historic, normal, and future periods are all returned in one table.

A \code{data.table} or SpatVector with downscaled climate variables. If \code{gcm} is NULL,
this is just the downscaled \code{normal} at point locations. If \code{gcm} is provided,
this returns a downscaled dataset for each point location, general circulation
model (GCM), shared socioeconomic pathway (SSP), run and period.
}
\description{
\code{climr_downscale} downscales and calculates climate variables for points of interest.

\code{downscale} downscales target rasters to points of interest.
}
\details{
Additional details... TODO.

Additional details... TODO.
}
\examples{
{
library(data.table)
library(terra)
set.seed(123)
dbCon <- data_connect()
xyz <- data.frame(lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10))

## get bounding box based on input points
thebb <- get_bb(xyz)
historic <- historic_input(dbCon, thebb, period = "2001_2020")
plot(historic[[1]][[2]])

## provide or create a long, lat, elev, and optionally id, dataframe - usually read from csv file
in_xyz <- data.frame(Long = c(-127.70521, -127.62279, -127.56235, -127.7162,
                              -127.18585, -127.1254, -126.94957, -126.95507),
                     Lat = c(55.3557, 55.38847, 55.28537, 55.25721, 54.88135, 54.65636, 54.6913, 54.61025),
                     Elev = c(291L, 296L, 626L, 377L, 424L, 591L, 723L, 633L),
                     Zone = c(rep("CWH",3), rep("CDF",5)),
                     Subzone = c("vm1","vm2","vs1",rep("mm",3),"dk","dc"))

## historic observational time series
climate_norms_hist <- climr_downscale(xyz = in_xyz, which_normal = "auto",
                                      return_normal = TRUE,
                                      historic_period = "2001_2020",
                                      vars = c("PPT","CMD","Tave07"),
                                      out_spatial = TRUE, plot = "PPT") ##specify desired variables to plot

## as a data.table
climate_norms_hist <- climr_downscale(xyz = in_xyz, which_normal = "auto",
                                      return_normal = TRUE,
                                      vars = cc("PPT","CMD","Tave07"),
                                      out_spatial = FALSE, plot = "PPT") ##specify desired variables to plot

## future projections
climate_norms_fut <- climr_downscale(xyz = in_xyz, which_normal = "auto",
                                     gcm_models = c("ACCESS-ESM1-5"),
                                     ssp = c("ssp370"),
                                     gcm_period = c("2021_2040"),
                                     #gcm_ts_years = 2020:2060,
                                     max_run = 3, #' we want 3 individual runs for each model
                                     vars = c("PPT","CMD","Tave07"))
}
\dontrun{
dbCon <- data_connect()
on.exit(try(pool::poolClose(dbCon)))
xyz <- data.frame(lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10))

## get bounding box based on input points
thebb <- get_bb(xyz)
normal <- normal_input(dbCon = dbCon, bbox = thebb, cache = TRUE)

## pick one GCM, one SSP and one period from the list of available options 
gcm <- gcm_input(dbCon, thebb, gcm = list_gcm()[3], list_ssp()[1], list_gcm_period()[2])

## notice coarseness of the data
terra::plot(gcm[[1]])

downscale(xyz, normal, gcm)
historic <- historic_input(dbCon, thebb)
terra::plot(historic[[1]])

downscale(xyz, normal, gcm = NULL, historic = historic, ppt_lr = FALSE)
}
}
\seealso{
\code{\link[=gcm_input]{gcm_input()}}, \code{\link[=historic_input]{historic_input()}}, \code{\link[=list_variables]{list_variables()}}
}
\author{
Kiri Daust
}
