% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R,
%   R/external_functions_old.R
\name{external_functions}
\alias{external_functions}
\alias{nc_get_variable_list}
\alias{nc_get_dim_bounds_var_list}
\alias{nc_get_climatology_bounds_var_list}
\title{Copy-pasted functions from other packages}
\usage{
nc_get_variable_list(f, min.dims = 1)

nc_get_dim_bounds_var_list(f, v = NULL)

nc_get_climatology_bounds_var_list(f)

nc_get_variable_list(f, min.dims = 1)

nc_get_dim_bounds_var_list(f, v = NULL)

nc_get_climatology_bounds_var_list(f)
}
\description{
Some single functions from packages are copied here to avoid having to import the library

All functions have permissive licenses from their cited packages

Included packages:
    ncdf4.helpers

NA
}
\author{
Brandon McNellis
}
