% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_db}
\alias{extract_db}
\title{Bilinear interpolation point extraction from raster bands}
\usage{
extract_db(dbCon, xyz, rastertbl, bands = NULL, nm = NULL, explain = FALSE)
}
\arguments{
\item{dbCon}{A postgres database connection.}

\item{xyz}{The name of the temporary point table in the database.}

\item{rastertbl}{The name of the raster table to extract from.}

\item{bands}{The index number of the bands to extract.}

\item{nm}{The names of the bands. Same length as bands.}

\item{explain}{Get the query execution profile for performance analysis.}
}
\value{
A data.frame one row per point id in xyz with an ID
column + one columns for each \code{bands} in \code{rastertbl} name \code{nm}.
}
\description{
Bilinear interpolation point extraction from raster bands
}
