% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.climr.R
\name{fit}
\alias{fit}
\title{Fit basic statistical models to climate data}
\usage{
fit(obj, data_type = c("yearly", "quarterly", "monthly"), fit_type = c("lm",
  "loess", "smooth.spline"))
}
\arguments{
\item{obj}{An object of class \code{climr} from \code{\link{load_clim}}}

\item{data_type}{The type of data to be analysed, either yearly, quarterly, or monthly}

\item{fit_type}{The type of model required, either linear regression (\code{lm}), loess, or smoothing spline (\code{smooth.spline})}
}
\value{
Returns a list of class \code{climr_fit} which includes the model details as well as the data set and fit type used
}
\description{
Fit basic statistical models to climate data
}
\examples{
ans1 = load_clim('SH')
ans2 = fit(ans1)
ans3 = fit(ans1, data_type = 'monthly', fit_type = 'smooth.spline')
ans4 = fit(ans1, data_type = 'quarterly', fit_type = 'loess')
}
\seealso{
\code{\link{load_clim}}, \code{\link{plot.climr_fit}}
}

