% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reml.R
\name{fit_reml}
\alias{fit_reml}
\title{Two-way Hierarchical Analysis of Variance}
\usage{
fit_reml(data, control = lme4::lmerControl(optimizer = "Nelder_Mead"),
  ...)
}
\arguments{
\item{data}{A dataframe inheriting from class \code{"cmpi5"}, containing
variables \code{index}, \code{GCM}, \code{RCP} and \code{run}.  See
\code{\link{cmip5_temp1}}.}

\item{control}{The argument \code{control} to
\code{\link[lme4]{lmerControl}}, normally, using the optimizer named
\code{"Nelder_Mead"}.}

\item{...}{Further arguments to be passed to \code{\link[lme4]{lmer}}.}
}
\value{
An object (a list) of class \code{climr_reml} and \code{climr}
  containing the following components.

  \item{call}{the call of \code{fit_reml}} \item{lmer_object}{the result of
  fitted model using the REML method from \code{\link[lme4]{lmer}}}
  \item{ests}{containing the estimated coefficients for \eqn{\mu} and the
  estimated standard deviations of random-effects factors} \item{ses}{the
  standard errors of estimators for fixed-effects and that of estimated
  standard deviations of random-effects factors} \item{ci}{containing the
  upper limits (i.e 2.5\%) and lower limits (i.e 97.5\% ) of the 95\% confidence
  intervals of estimators for fixed-effects and that of estimated standard
  deviations of random-effects factors}
}
\description{
Description fitting a random-effects ANOVA model and finding the standard
deviations of estimators by using the restricted maximum likelihood (REML)
estimation, following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Chandler
(2014)}
}
\details{
The data would fit in a two-way random-effects ANOVA model
  containing three factors which are GCM, RCP and the interaction between GCM
  and RCP. Those factors are all considered to be normally distributed random
  variables with mean zero and different standard deviation, like
  \eqn{\sigma_{GCM)},\sigma_{RCP},\sigma_{GCM:RCP}}{\sigma_(GCM),\sigma_(RCP),\sigma_(GCM:RCP)}.
  In addition, the intercept \eqn{\mu} of this model interprets the overall
  mean change in the index and the error term \eqn{\epsilon} are independent
  identically distributed random variables with mean zero and standard
  deviation \eqn{\sigma_{run}}{\sigma_(run)}.

  This function uses the restricted maximum likelihood (REML) method for
  estimating those estimators and their corresponding standard errors.
}
\examples{
library(lme4)
res <- fit_reml(cmip5_temp1)
}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying
  Sources of Uncertainty in Projections of Future Climate.
  \emph{Journal of Climate}, \strong{27}, 8793-8808.
  \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip5_temp1}} and \code{\link[climr]{cmip5_temp2}}
 for CMIP5 data.

\code{\link[lme4]{lmer}} for further arguments.

\code{\link[climr]{print.climr_reml}} for \code{print}.

\code{\link[climr]{summary.climr_reml}} for \code{summary}.

\code{\link[climr]{print.summary.climr_reml}} for \code{print.summary}.
}
