% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{gcm_ts_input}
\alias{gcm_ts_input}
\title{Create gcm timeseries input for \code{downscale} using data on Postgis database.}
\usage{
gcm_ts_input(
  dbCon,
  bbox = NULL,
  gcm = list_gcm_ts(),
  ssp = list_ssp(),
  years = list_years_ts(),
  max_run = 0L,
  cache = TRUE
)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{Numeric vector of length 4 giving bounding box of study region, create by \code{get_bb()}}

\item{gcm}{A character vector. Label of the global circulation models to use.
Can be obtained from \code{list_gcm()}. Default to \code{list_gcm()}.}

\item{ssp}{A character vector. Label of the shared socioeconomic pathways to use.
Can be obtained from \code{list_ssp()}. Default to \code{list_ssp()}.}

\item{years}{Numeric or character vector in in \code{2020:2100}}

\item{max_run}{An integer. Maximum number of model runs to include.
A value of 0 is \code{ensembleMean} only. Runs are included in the order they are found in the
models data untile \code{max_run} is reached. Default to 0L.}

\item{cache}{Logical specifying whether to cache new data locally or no. Default \code{TRUE}}
}
\value{
An object to use with \code{downscale}. A list of \code{SpatRaster} with, possibly, multiple layers.
}
\description{
Create gcm timeseries input for \code{downscale} using data on Postgis database.
}
