% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_elev_raster}
\alias{get_elev_raster}
\title{Create an elevation raster from a template raster}
\usage{
get_elev_raster(r, elev, out = NULL, ...)
}
\arguments{
\item{r}{A SpatRaster. The template.}

\item{elev}{An elevation raster such as (northamerica_elevation_cec_2023.tif) from
http://www.cec.org/files/atlas_layers/0_reference/0_03_elevation/elevation_tif.zip}

\item{out}{A file path out for \code{terra::writeRaster}. Default NULL.}

\item{...}{Additional parameters to \code{terra::writeRaster}.}
}
\description{
Create an elevation raster from a template raster
}
\examples{
r <- terra::rast(nrows=108, ncols=21, xmin=0, xmax=10)
elev <- system.file("ex/elev.tif", package="terra") |> terra::rast()
get_elev_raster(r, elev, out = "elev.tif", gdal = "PREDICTOR=2", datatype="FLT4S", overwrite = TRUE)

}
