% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latitude_raster}
\alias{get_latitude_raster}
\title{Create a latitude raster}
\usage{
get_latitude_raster(r, out = NULL, ...)
}
\arguments{
\item{r}{A SpatRaster.}

\item{out}{A file path out for \code{terra::writeRaster}. Default NULL.}

\item{...}{Additional parameters to \code{terra::writeRaster}.}
}
\description{
Create a latitude raster
}
\examples{
get_latitude_raster("Tmin_01.tif", out = "alt.tif", gdal = "PREDICTOR=2", datatype="FLT4S", overwrite = TRUE)
}
