% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\name{get_station_data}
\alias{get_station_data}
\title{Retrieval of station data near points or in bounding box}
\usage{
get_station_data(
  x,
  dataset = "AHCCD",
  k = 1,
  date_range = c("2001-01-01", "2005-12-31"),
  vars = sort(sprintf(c("PPT_\%02d", "Tmax_\%02d", "Tmin_\%02d"), sort(rep(1:12, 3)))),
  only_complete = FALSE
)
}
\arguments{
\item{x}{Either a data.frame containing point locations with columns \code{id}, \code{lat}, \code{lon} OR a numeric vector of length four
representing a bounding box to select stations within. This can be created using \code{get_bb()} or it can be a SpatExtent}

\item{dataset}{Character. Which station dataset to use? Current options are \code{AHCCD} or \code{SCDNA}. Default "AHCCD"}

\item{k}{Integer. Number of stations to return for each point. Only used if \code{x} is a data.frame.}

\item{date_range}{Character vector of length 2. Date range to retrieve data for. Default is \code{c("2001-01-01", "2005-12-31")}.}

\item{vars}{character. A vector of climate variables to compute. Supported variables
can be obtained with \code{\link[=list_vars]{list_vars()}}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}

\item{only_complete}{Logical. If TRUE, only stations with complete Tmin, Tmax, PPT are returned. Default \code{FALSE}.}
}
\value{
data.table containing requested variables for stations near points or within bounding box.
If \code{x} is points, the distance (in meters) and rank of each station relative to each input point will also be returned.
}
\description{
This function aims to facilitate the retrieval of station data. The function allows specification
by area, dataset, date range, and completeness. For stations with complete observations (i.e., Tmin, Tmax, PPT),
the function calculates the standard suite of derived variables in climr. If \code{x} is a data.frame of point locations,
the function will return the \code{k} nearest stations to each input point. If \code{x} is a bounding box (numeric vector or
SpatExtent), the function returns all stations withing the extent.
}
