% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Retrieve station metadata for given stations}
\usage{
get_station_metadata(station_ids)
}
\arguments{
\item{station_ids}{character vector of station ids}
}
\value{
data.table containing lon, lat, elev, dataset, and province for each requested station
}
\description{
Retrieve station metadata for given stations
}
