% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{getIRL_dGSAT}
\alias{getIRL_dGSAT}
\title{getIRL_dGSAT}
\usage{
getIRL_dGSAT(
  scenario_name,
  rcp_path,
  ecs = 3,
  kappa = 2.3,
  alpha_a = 1,
  alpha_v = 1,
  C_0 = 276.0897,
  beta = 0.36,
  Q_10 = 2,
  M0 = 653
)
}
\arguments{
\item{scenario_name}{national scenario}

\item{rcp_path}{pathway (RCP or SSP)}

\item{ecs}{equilibrium climate sensitivity}

\item{kappa}{vertical ocean diffusivity (default 2.3 cm2/s)}

\item{alpha_a}{anthropogenic aerosol scaling (unitless, default value unity)}

\item{alpha_v}{volcanic aerosol scaling (unitless, default value unity)}

\item{C_0}{preindustrial co2 concentration (ppmv)}

\item{beta}{carbon ferilisation}

\item{Q_10}{soil carbon feedback}

\item{M0}{pre-industrial methane concentration (ppbv)}
}
\value{
tibble
}
\description{
Ireland contribution to GSAT (global-mean surface air temperature)
wrapper function based on getIRLContrib
}
