% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{getIRLContrib}
\alias{getIRLContrib}
\title{getIRLContrib}
\usage{
getIRLContrib(scenario_name, rcp_path, ecs)
}
\arguments{
\item{scenario_name}{name of scenario from scenarios A, B, C, D, E}

\item{rcp_path}{names of RCP pathway currently RCP29,RCP45,RCP60, RCP85}

\item{ecs}{value for equilibrium climate sensitivity (typical 3C)}
}
\value{
a tibble year Tgav rcp scenarion
}
\description{
compute the Irish contribution to global warming based on historical data in ghg_irl
}
\examples{
getIRLContrib("A","rcp85",2.9)

}
