% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_irl.R
\docType{data}
\name{ghg_irl}
\alias{ghg_irl}
\title{Estimated well mixed greenhouse gas emissions Ireland 1745-2018}
\format{
A data frame with 1100 rows and 3 variables:
\describe{
\item{year}{price, in US dollars}
\item{ktCO2e}{weight of the diamond, in carats}
\item{gas}{gas type. lulucf is land-co2}
}
}
\source{
1990-2018 \url{http://www.epa.ie/} \url{http://www.cso.ie/}
}
\usage{
ghg_irl
}
\description{
A dataset containing annual emissions of ktCO2-eq of CO2 (fossil-co2), CH4, N2O and LULUCF (CO2).
}
\keyword{datasets}
