% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fit.climr.R
\name{gp_fit}
\alias{gp_fit}
\title{Fit basic statistical models to climate data}
\usage{
gp_fit(obj, type = c("Nelder-Mead", "BFGS", "SANN", "Brent"))
}
\arguments{
\item{obj}{An object of class \code{climr} from \code{\link{load_clim}}}

\item{type}{The type of method required, either Nelder-Mead (\code{Nelder-Mead}), Broyden-Fletcher-Goldfarb-Shanno (\code{BFGS}) , Simulated Annealing (\code{SANN}) or Brent method (\code{Brent})}
}
\value{
Returns a list of class \code{climr_gp_fit} which includes the model details as well as the data set and method type used
}
\description{
Fitting the model for the Gaussian Regression process and returning a list of values for plotting.
}
\examples{
data1 = load_clim('NH')
data2 = gp_fit(data1,type="BFGS")
}
\seealso{
\code{\link{load_clim}}, \code{\link{plot.climr_gp_fit}}
}
