% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMET.R
\name{gridMET}
\alias{gridMET}
\alias{aggregate_gridMET_by_year}
\alias{subset_gridMET_by_latlon}
\alias{collapse_gridMET_subset}
\title{Functions for gridMET/MACA manipulation}
\usage{
aggregate_gridMET_by_year(data_dir, vars, yrs, FUN)

subset_gridMET_by_latlon(coords, by, data_dir, vars, yrs)

collapse_gridMET_subset(subset_array, coords, by = "PLOT", FUN = mean)
}
\description{
NA
}
\author{
Brandon McNellis
}
