% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwps.R
\docType{data}
\name{gwps}
\alias{gwps}
\title{Global Warming Potentials}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{gas}{wmghg}
\item{gwp}{GWP100}
}
}
\source{
IPCC AR4
}
\usage{
gwps
}
\description{
Accounting use only
}
\keyword{datasets}
