% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector_esm_fits.R
\docType{data}
\name{hector_esm_fits}
\alias{hector_esm_fits}
\title{Hector fit parameters from Dorheim et al}
\format{
A data frame with 6 rows and 8 variables:
\describe{
\item{ESM}{ESM model code}
\item{alpha_a}{anthropgenic aerosol scaling factor}
\item{alpha_v}{volcano aerosol scaling factor}
\item{ecs}{equilibrium climate sensitivity}
\item{kappa}{vertical ocean diffusivity}
\item{C_0}{preindustrial co2}
\item{beta}{carbon fertilisation}
\item{Q_10}{Q10}
}
}
\source{
\url{https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2019EA000980}
}
\usage{
hector_esm_fits
}
\description{
6 emissions driven ESMs, partially penalised 7-parameter regression fits
}
\keyword{datasets}
