% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_input}
\alias{historic_input}
\title{Create historic input for \code{downscale}.}
\usage{
historic_input(dbCon, bbox = NULL, period = list_historic(), cache = TRUE)
}
\arguments{
\item{period}{A character vector. Label of the period to use.
Can be obtained from \code{list_period()}. Default to \code{list_period()}.}
}
\value{
An object to use with \code{downscale}. A \code{SpatRaster} with, possibly, multiple layers.
}
\description{
Create historic input for \code{downscale}.
}
