% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_input_ts}
\alias{historic_input_ts}
\title{Create historic timeseries input for \code{downscale}.}
\usage{
historic_input_ts(dbCon, bbox = NULL, years = 2010:2022, cache = TRUE)
}
\arguments{
\item{dbCon}{Database connection}

\item{bbox}{Bounding box of data, from \code{get_bb}}

\item{years}{Years to retrieve timeseries for, in \code{1902:2022}. Default \code{2010:2022}}
}
\value{
An object to use with \code{downscale}. A \code{SpatRaster} with, possibly, multiple layers.
}
\description{
Create historic timeseries input for \code{downscale}.
}
