% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{is_in_bbox}
\alias{is_in_bbox}
\title{TODO add documentation here}
\usage{
is_in_bbox(newbb, oldbb)
}
\arguments{
\item{newbb}{}

\item{oldbb}{}
}
\value{
logical
}
\description{
TODO add documentation here
}
