% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate}
\alias{lapse_rate}
\title{Lapse rate computation}
\usage{
lapse_rate(normal, dem, NA_replace = TRUE, nthread = 1L, rasterize = TRUE)
}
\arguments{
\item{normal}{a \code{SpatRaster} stack. Normal rasters to compute lapse rates for. Build with this package functions.}

\item{dem}{\code{SpatRaster}. A digital elevation model raster. Only needed if \code{normal}
is a \code{SpatRaster}. Defaults to \code{NULL}.}

\item{NA_replace}{logical. Should NA lapse rate results be replaced by zeros. Default to TRUE.}

\item{nthread}{integer. Number of parallel threads to use to compute lapse rates.}

\item{rasterize}{logical. Return an object of the same class category as normal with the same extent?
Default to \code{TRUE}.}
}
\value{
\code{SpatRaster} of lapse rate values.
}
\description{
Lapse rate computation
}
\details{
Formulas
Simple linear regression without the intercept term
beta_coef = sum(xy) / sum(x²)
mss = sum(x * beta_coef)², sum of squared fitted values
rss = sum(ε²), sum of squared (y minus fitted), sum of absolute errors
R² = mss / (mss + rss)
Lapse rate = beta_coef * R²
}
