% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate_mem}
\alias{lapse_rate_mem}
\title{Memoised version}
\usage{
lapse_rate_mem(
  target,
  NA_replace = TRUE,
  use_parallel = TRUE,
  rasterize = TRUE
)
}
\arguments{
\item{target}{Target rasters to compute lapse rates for. Build with this package functions.}

\item{NA_replace}{A boolean. Should NA lapse rate results be replaced by zeros. Default to TRUE.}

\item{use_parallel}{A boolean. Should parallel::mclapply be used. Default to TRUE.}

\item{rasterize}{Return an object of the same class category as target with the same extend.}
}
\description{
Memoised version
}
