% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rates}
\alias{lapse_rates}
\title{Compute and cache lapse rates for later use}
\usage{
lapse_rates(
  normal,
  dem,
  NA_replace = TRUE,
  use_parallel = TRUE,
  rasterize = TRUE
)
}
\arguments{
\item{normal}{Normal rasters to compute lapse rates for. Build with this package functions.}

\item{dem}{A digital elevation model with matching normal extent.}

\item{NA_replace}{A boolean. Should NA lapse rate results be replaced by zeros. Default to TRUE.}

\item{use_parallel}{A boolean. Should parallel::mclapply be used. Default to TRUE.}

\item{rasterize}{Return an object of the same class category as normal with the same extend.}
}
\description{
Compute and cache lapse rates for later use
}
