% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_data}
\alias{list_data}
\title{List package local cache files}
\usage{
list_data(subdirectory)
}
\arguments{
\item{subdirectory}{A character. A subdirectory of \code{data_path()}. Restrict listing to only
this particular subdirectory. Use \code{getOption("climr.gcm.path")} or
\code{getOption("climr.normal.path")}.}
}
\description{
List package local cache files
}
