% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{list_parse}
\alias{list_parse}
\title{Read and parse gcm models csv files}
\usage{
list_parse(gcm, col_num = 1)
}
\arguments{
\item{gcm}{An optional character vector. Limit list to provided global circulation models.}

\item{col_num}{An integer vector.}
}
\value{
A character vector of unique values.
}
\description{
Read and parse gcm models csv files
}
