% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{list_period}
\alias{list_period}
\title{List available period}
\usage{
list_period(gcm)
}
\arguments{
\item{gcm}{An optional character vector. Limit list to provided global circulation models.}
}
\description{
List available period
}
