% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{list_unique}
\alias{list_unique}
\title{Read and parse gcm models csv files}
\usage{
list_unique(files, col_num)
}
\arguments{
\item{files}{character. A vector of file paths.}

\item{col_num}{integer. Vector of indices of elements to retrieve in label. Label is split
by "_" before processing.}
}
\value{
A character vector of unique values.
}
\description{
Read and parse gcm models csv files
}
