% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{list_variables}
\alias{list_variables}
\title{List climate variables}
\usage{
list_variables(include_default = TRUE)
}
\arguments{
\item{include_default}{A boolean. Should Tmin, Tmax and PPT be included? Default to TRUE.}
}
\description{
List climate variables
}
