% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clim.R
\name{load_clim}
\alias{load_clim}
\title{Load in climate data from NASA}
\usage{
load_clim(type = c("GLB", "NH", "SH"))
}
\arguments{
\item{type}{Either GLB, NH or SH for global, northern or southern hemisphere temperature anomalies respectively}
}
\value{
A list of \code{\link[tibble]{tibble}}s which contain the yearly, quarterly, and monthly values for each time series respectively, as well as which series was obtained (i.e. GLB, NH or SH)
}
\description{
Load in climate data from NASA
}
\examples{
data = load_clim(type = 'SH')
}
\seealso{
\code{\link{fit}}, \code{\link{plot.climr_fit}}
}

