% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_climr.R
\name{load_climr}
\alias{load_climr}
\title{Load in climate data from NASA}
\usage{
load_climr(type = c("GLB", "NH", "SH"))
}
\arguments{
\item{type}{Either \code{"GLB"}, \code{"NH"}, or \code{"SH"} for global, northern,
or southern hemisphere temperature anomalies respectively. Defaults to \code{"GLB"}.}
}
\value{
An object of class \code{"climr"} which is a list of \code{\link[tibble]{tibble}}s which
contain the yearly, quarterly, and monthly values for each time series respectively.
}
\description{
Loads in global or hemispheric data from NASA at different intervals.
}
\note{
A dedicated \code{\link{fit}} function is provided for objects of class \code{"climr"}.
}
\examples{
dat <- load_climr(type = "SH")
}
\seealso{
\code{\link{fit}}, \code{\link{plot.climr_fit}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@mu.ie}>
}
