% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logVars.R
\name{logVars}
\alias{logVars}
\title{Log-transform climate variables}
\usage{
logVars(
  dat,
  elements = c("AHM", "DD", "Eref", "FFP", "NFFD", "PAS", "PPT", "SHM", "CMD"),
  base = exp(1),
  add.fields = FALSE,
  zero_adjust = TRUE
)
}
\arguments{
\item{dat}{a \code{data.table} with columns of climate variables including
the selected climate \code{elements}.}

\item{elements}{character. Climate elements to log-transform.}

\item{base}{numeric. Logarithmic base.}

\item{add.fields}{logical. If \code{TRUE}, the new logged variables are added to \code{dat}
(TRUE). Otherwise, original column values are replaced with the logs (FALSE).}

\item{zero_adjust}{logical. If \code{TRUE} adjusts zeroes in raw data as:
\eqn{base^{\log_base{x_min} - 1}}.
where \eqn{x_min} is the minimum non-zero, non-NA value.}
}
\value{
\code{data.table}
}
\description{
This function performs logarithmic transformation on climr output to a user-specified base.
This function currently only works with climr output in data.table format (i.e., not with raster output).
}
\details{
In climate analysis it is often preferable to log-transform ratio variables, i.e., variables such as precipitation and
degree-days that have a lower limit of zero and for which proportions are meaningful.

Climate elements include all related time variables, hence all column names that
partially match strings in \code{elements} will be log-transformed. For example, specifying
\code{elements = "PPT"} will #'   result in log-transformation of \code{c("PPT", "PPT_01", "PPT_02", ... , "PPT_at")}.

By default, zeroes are assigned a log value that is one order of magnitude less than the minimum positive value
of the untransformed variable.
}
