% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{make_raster}
\alias{make_raster}
\title{Make raster from a boundary}
\usage{
make_raster(boundary, conn, rastque, nameque, projID, bands)
}
\arguments{
\item{boundary}{}

\item{conn}{}

\item{rastque}{}

\item{nameque}{}

\item{projID}{}

\item{bands}{}
}
\value{
a SpatRaster
}
\description{
Used internally to access the PostGRS database and
create a SpatRaster using a given spatial boundary
}
