% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{std_err}
\alias{trimws_df}
\alias{fac_to_char}
\alias{mode0}
\alias{inv_logit}
\title{Miscellaneous supporting functions}
\usage{
std_err(x, na.rm = F)

trimws_df(x, stringsAsFactors = F)

fac_to_char(x)

mode0(x)

inv_logit(x)
}
\description{
NA
}
\author{
Brandon McNellis
}
