% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTBS.R
\name{MTBS}
\alias{MTBS}
\alias{get_MTBS}
\alias{filter_MTBS_by_Cleland}
\alias{filter_MTBS_by_year}
\title{Functions for MTBS manipulation}
\usage{
get_MTBS(dir)

filter_MTBS_by_Cleland(shp, sections)

filter_MTBS_by_year(shp, years)
}
\description{
NA
}
\author{
Brandon McNellis
}
