% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_changes.R
\docType{data}
\name{name_changes}
\alias{name_changes}
\title{Changes to naming conventions applied in Version 0.1.0}
\format{
A \code{data.table} with columns:
\describe{
\item{type}{The type of object that was renamed}
\item{function}{The function whose contents were renamed, if applicable}
\item{name_old}{The old name, pre-v0.1.0}
\item{name_new}{The new name, as of v0.1.0}
}
}
\usage{
name_changes
}
\description{
A table of correspondence between old and new names for functions, parameters, and options.
For changes in climate variable names, see \code{data(variables)}.
}
\keyword{internal}
