% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Create normal baseline}
\usage{
normal(normal = list_normal()[1], dem = list_dem()[1])
}
\arguments{
\item{normal}{A character. Name of the normal baseline to use. Can be obtained from
\code{list_normal()}. Default to \code{list_normal()[1]}.}

\item{dem}{A character. Name of the digital elevation model to use. Must have the same
extent as the baseline normal. Can be obtained from \code{list_dem()}. Default to \code{list_dem()[1]}.}
}
\value{
A normal baseline to use with \code{downscale}. A \code{SpatRaster} with a \code{dem} attribute.
}
\description{
Create normal baseline
}
