% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{packRasters}
\alias{packRasters}
\title{Pack rasters for parallel computing}
\usage{
packRasters(ras)
}
\arguments{
\item{ras}{a SpatRaster or list of SpatRasters, or \code{NULL}}
}
\value{
\code{NULL}, a packed SpatRaster or list of packed SpatRaters
}
\description{
Pack rasters for parallel computing
}
