% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{List of Shared Socioeconomic Pathways (and RCPs)}
\format{
A vector of valid SSPs or RCPs
\describe{
\item{}{list of names}
}
}
\source{

}
\usage{
pathways
}
\description{
List of Shared Socioeconomic Pathways (and RCPs)
}
\keyword{datasets}
