% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{pgGetTerra}
\alias{pgGetTerra}
\title{Download raster with bounding box from PostGIS}
\usage{
pgGetTerra(conn, name, tile, rast = "rast", bands = 37:73, boundary = NULL)
}
\arguments{
\item{conn}{a DBI or RPostgres connection object}

\item{name}{character. Table name in database}

\item{tile}{TODO}

\item{rast}{character. Name of column which stores raster data.
Defaults to "rast"}

\item{bands}{Which raster bands to return. Default 37:73.}

\item{boundary}{Numeric vector of length 4 (ymax,ymin,xmax,xmin). Defaults to \code{NULL}.}
}
\value{
A SpatRaster
}
\description{
Download raster with bounding box from PostGIS
}
