% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.climr_bayes.R
\name{plot.climr_bayes}
\alias{plot.climr_bayes}
\title{Descriptive plot for Bayesian Analysis}
\usage{
\method{plot}{climr_bayes}(data, which = 1, super_pop = TRUE, ...)
}
\arguments{
\item{data}{An object of class \code{"climr_bayes"} form
\code{\link[climr]{fit_bayes}}, containing the \code{brm_object} that is a
result of Bayesian inference (i.e. \code{\link[brms]{brm}}).}

\item{which}{The choice of types of plots (default to be 1). See also in
Details.}

\item{super_pop}{The choice of the types of standard deviation. As
\code{super_pop = TRUE}, it shows the super-population standard deviation.
Otherwise, it shows the finite-population standard deviation.}

\item{...}{Further arguments to be passed to
\code{\link[bayesplot]{mcmc_intervals}}, \code{\link[graphics]{hist}},
\code{\link[graphics]{curve}} and \code{\link[graphics]{legend}}.}

\item{prob}{The probability mass to include in the inner interval for
\code{\link[bayesplot]{mcmc_intervals}}, and the default is 0.5 (i.e.50\%
confidence interval)}

\item{prob_outer}{The probability mass to include in the outer intervalfor
\code{\link[bayesplot]{mcmc_intervals}}, and the default is 0.95 (i.e.95\%
confidence interval)}

\item{point_est}{The point estimate to show. Either "median" (the default),
"mean", or "none".}

\item{A}{Setting for the parameters of the half-Cauchy prior distributions of
standard deviations of GCM, RCP and their interaction, i.e. the group-level
effects and of the error standard deviation, that is, the standard
deviation over runs for a given GCM-RCP combination.(defaults to be 0.5)}

\item{col_bar}{The color of the bar in the histogram.(more information in
\code{\link{par}})}

\item{col_line}{The color of the line (i.e. curve) in the histgram.(more
information in \code{\link{par}})}

\item{legend_pos}{Position of the legend, nominally,is in \code{"center"}.}

\item{lwd}{The line width, a positive number, defaulting to 2. (more
information in \code{\link{par}})}
}
\value{
Nothing is returned.
}
\description{
\code{\link[graphics]{plot}} method for an object of class
\code{"climr_bayes"},following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Richard
(2014)}. Here is two types of plots that are summaries of the posterior
distributions for the super-population standard deviations of the four
random effects (\code{which = 1, super_pop = TRUE}), summaries of the posterior
distributions for the finite-population standard deviations of the four
random effects (\code{which = 1, super_pop = FALSE}) and their posterior distrbutions
based on the half-Cauchy prior distrbutions (\code{which = 2, super_pop = TRUE}).
}
\details{
This function is only applicable for the Bayesian result from CMIP5
  data whose class is \code{"climr_bayes"} from
  \code{\link[climr]{fit_bayes}}, containing the \code{brm_object} that is
  the output directly from the \code{\link[brms]{brm}}. Here are two types of
  plots in the function.

  The first one is the plot from \code{\link[bayesplot]{mcmc_intervals}} as
  \code{which = 1}. This plot creates confidence intervals computed from
  posterior distribution. For CMIP5, there are four lines interpreting
  \code{sd_GCM_Intercept}, \code{sd_RCP_Intercept},
  \code{sd_GCM:RCP_Intercept} and \code{sigma}, corresponding to
  \eqn{\sigma_{GCM}}{\sigma_GCM}, \eqn{\sigma_{RCP}}{\sigma_RCP},
  \eqn{\sigma_{GCM:RCP}}{\sigma_GCM:RCP} and \eqn{\sigma_{run}}{\sigma_run},
  respectively. However, for CMIP6, there are four lines interpreting
  \code{sd_GCM_Intercept}, \code{sd_SSP_Intercept},
  \code{sd_GCM:SSP_Intercept} and \code{sigma}, corresponding to
  \eqn{\sigma_{GCM}}{\sigma_GCM}, \eqn{\sigma_{SSP}}{\sigma_SSP},
  \eqn{\sigma_{GCM:SSP}}{\sigma_GCM:SSP} and \eqn{\sigma_{run}}{\sigma_run},
  respectively. The dot represents the \code{point_est} and the default is
  showing the median position. The thick lines are corresponding to the 50\%
  confidence interval which could be setted in \code{prob} and the thin lines
  are corresponding to the 95\% confidence interval which could also be setted
  in \code{prob_outer}. But there is one thing that the value of \code{prob}
  should not close to that of \code{prob_outer}. At the same time with
  \code{which = 1}, if \code{super_pop = TRUE}, those values are corresponding
  to the super-population standard deviations. Otherwise, if \code{super_pop = FALSE},
  those values are from the finite-population standard deviations.

  If choosing \code{which = 2}, it would show the plot from
  \code{\link[graphics]{hist}} creating a histogram of posterior distributions
  of \eqn{\sigma_{GCM}}{\sigma_GCM}, \eqn{\sigma_{RCP}}{\sigma_RCP},
  \eqn{\sigma_{GCM:RCP}}{\sigma_GCM:RCP} and \eqn{\sigma_{run}}{\sigma_run} for
  CMIP5 or of \eqn{\sigma_{GCM}}{\sigma_GCM}, \eqn{\sigma_{SSP}}{\sigma_SSP},
  \eqn{\sigma_{GCM:SSP}}{\sigma_GCM:SSP} and \eqn{\sigma_{run}}{\sigma_run}
  for CMIP6, both based on the half-Cauchy prior distribution with parameter
  \code{A} that is interpreted by the curve from \code{\link[graphics]{curve}}.
  From this plot, we could see the fitness of prior distrbuiton with the value
  of \code{A} to the posterior distribution. With a suitable value of \code{A}
  in the half-Cauchy prior distrbution, there is a hight probability on the
  realistic range. But the prior has a "long tail" for preventing an undue
  influence from the prior if the data suggest a larger than anticipated value
  of \eqn{\sigma}. Besides, the \code{which = 2} is only working for
  \code{super_pop = TRUE}. As the \code{which = 2} and \code{super_pop = FALSE}
  are choosing at the same time, it would stop.

  The relatice mathematics in this
  pacakge are seen in \code{vignette("climr", package = "climr")} for an
  overview of the package.
}
\examples{
x <- fit_bayes(cmip5_temp1)
y <- fit_bayes(cmip6_pr)
### Plot method [1] for posterior distribution with super-population standard deviation ###
plot(x)
plot(y)
### Plot method [1] for posterior distribution with finite-population standard deviation###
plot(x, super_pop = FALSE)
plot(y, super_pop = FALSE)
### Plot method [2] for prior distribution ###
plot(x, which = 2)
plot(y, which = 2)

}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying Sources of
  Uncertainty in Projections of Future Climate. \emph{Journal of Climate},
  \strong{27}, 8793-8808. \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip5_temp1}} and
  \code{\link[climr]{cmip5_temp2}} for CMIP5 data.

\code{\link[climr]{cmip6_pr}} and \code{\link[climr]{cmip6_tas}} for
  CMIP6 data.

\code{\link[climr]{fit_bayes}} for operation of bayesian inference.

\code{\link[bayesplot]{mcmc_intervals}} for bayesplot of the posterior
distrbutions
}
