% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.climr_gp_fit.R
\name{plot.climr_gp_fit}
\alias{plot.climr_gp_fit}
\title{Plot statistical model output to the climate data}
\usage{
\method{plot}{climr_gp_fit}(x, ...)
}
\arguments{
\item{x}{Output from the \code{\link{gp_fit}} function}

\item{...}{Other arguments to plot (not currently implemented)}
}
\value{
Displaying the Gaussian process regression plot of the NASA dataset for the years running from 1880-2020.
}
\description{
Plot climr_gp_fit output for the Gaussian Regression process
}
\examples{
data1 = load_clim('NH')
data2 = gp_fit(data1,type="BFGS")
plot(data2)

}
\seealso{
\code{\link{load_clim}}, \code{\link{fit}}, \code{\link{gp_fit}}
}
