% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cmip5.R
\name{plot.cmip5}
\alias{plot.cmip5}
\title{Descriptive plot for CMIP5 data}
\usage{
\method{plot}{cmip5}(data, legend_pos = "topright",
  legend_text = levels(data$RCP), ...)
}
\arguments{
\item{data}{An object of class \code{"cmip5"}, a result of a call to
\code{\link{nnt}}.}

\item{legend_pos}{Position of the legend, nominally,is in \code{"topright"}.}

\item{legend_text}{Text of the legend, nominally, be in the order of
\code{"RCP2.6", "RCP4.5", "RCP6.0", "RCP8.5"}.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}} or
\code{\link[graphics]{legend}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an object of class \code{"cmip5"},following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Richard
(2014)}
}
\details{
This function is only applicable for the CMIP5 data containing the
  index, GCM and RCP; Users could adjust the plot whatever they want, like
  the \code{pch} and \code{col}. Although there are some arguments to avoid
  the systemic misunderstanding between the \code{\link[graphics]{plot}} and
  \code{\link[graphics]{legend}}, it is still possible to create warning
  messages. But those could be eschewed by careful input.

  If users do not have any other further arguments, this function could also
  create a satisfied plot for them with classified colors and points.

  Colours and types of the points can be set sing an argument \code{col} and
  \code{pch}.If a variable is wrapped then the default plotting limits are
  set using the corresponding values in \code{data$RCP}.
}
\examples{
plot(cmip5_temp1)

}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying
  Sources of Uncertainty in Projections of Future Climate.
  \emph{Journal of Climate}, \strong{27}, 8793-8808.
  \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip5_temp1}} and \code{\link[climr]{cmip5_temp2}}
 for CMIP5 data.
}
