% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cmip6.R
\name{plot.cmip6}
\alias{plot.cmip6}
\title{Descriptive plot for CMIP6 data}
\usage{
\method{plot}{cmip6}(x, region = 0, period = 1,
  legend_pos = "topright", legend_text = levels(data$SSP), ...)
}
\arguments{
\item{region}{Region of the world: 0 for global, 1-22 for regions.(default 0)}

\item{period}{Time horizon: 1 for mid 21st century (2020-2049); 2 for late
21st century (2070-2099).}

\item{legend_pos}{Position of the legend, nominally,is in \code{"topright"}.}

\item{legend_text}{Text of the legend, nominally, be in the order of
\code{"ssp119", "ssp126", "ssp245", "ssp370", "ssp434", "ssp460",
"ssp534-over", "ssp585"}.}

\item{...}{Further arguments to be passed to \code{\link[graphics]{plot}} or
\code{\link[graphics]{legend}}.}

\item{data}{An object of class \code{"cmip6"}, a result of a call to
\code{\link{nnt}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for an object of class \code{"cmip6"},following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Richard
(2014)}
}
\details{
This function is only applicable for the CMIP6 data containing the
  index, GCM and SSP; Users could adjust the plot whatever they want, like
  the \code{pch} and \code{col}. Although there are some arguments to avoid
  the systemic misunderstanding between the \code{\link[graphics]{plot}} and
  \code{\link[graphics]{legend}}, it is still possible to create warning
  messages. But those could be eschewed by careful input.

  If users do not have any other further arguments, this function could also
  create a satisfied plot for them with classified colors and points.

  Colours and types of the points can be set sing an argument \code{col} and
  \code{pch}.If a variable is wrapped then the default plotting limits are
  set using the corresponding values in \code{data$SSP}.
}
\examples{
plot(cmip6_pr)

}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying Sources of
  Uncertainty in Projections of Future Climate. \emph{Journal of Climate},
  \strong{27}, 8793-8808. \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip6_pr}} and \code{\link[climr]{cmip6_tas}} for
  CMIP6 data.
}
