% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivariate_input.R
\name{plot_bivariate_input}
\alias{plot_bivariate_input}
\title{Input data for the bivariate climate change plot}
\usage{
plot_bivariate_input(
  xyz,
  obs_period = list_obs_periods()[1],
  gcms = list_gcms(),
  ssps = list_ssps()[1, 2, 3],
  gcm_periods = list_gcm_periods(),
  max_run = 10,
  vars = list_vars()
)
}
\arguments{
\item{xyz}{a \code{terra::SpatRaster} with a single layer containing elevation values in metres, or a \code{data.frame} with the following columns
"long", "lat", "elev", and a unique "id". Any extra columns will be ignored and not output.}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}.}

\item{max_run}{integer. Maximum number of model runs to include, not including the ensemble mean.
Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{vars}{character. A vector of climate variables to compute. Supported variables
can be obtained with \code{\link[=list_vars]{list_vars()}}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}
}
\value{
\code{data.table} of average downscaled climate variables for all locations.
}
\description{
Input data for the \code{\link[=plot_bivariate]{plot_bivariate()}} function. Since these inputs are time-consuming to
generate, the purpose of conducting the generation of the input table in a separate function is
to allow users to make multiple calls to \code{\link[=plot_bivariate]{plot_bivariate()}} (e.g., for comparing different
climate variables) without needing to generate the inputs each time.
}
\details{
This function generates standardized inputs for one or multiple locations at any spatial scale.
}
\examples{
if (FALSE) {
  # data frame of arbitrary points
  my_points <- data.frame(
    lon = c(-127.7300, -127.7500),
    lat = c(55.34114, 55.25),
    elev = c(711, 500),
    id = 1:2
  )

  # generate the input data
  my_data <- plot_bivariate_input(my_points)

  # use the input to create a plot
  plot_bivariate(my_data, xvar = "MAT", yvar = "PAS_an")
}
#'
}
