% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_diagram.R
\name{plot_climate_diagram}
\alias{plot_climate_diagram}
\title{Plot Climate Diagram}
\usage{
plot_climate_diagram(
  xyz,
  gcms = list_gcms()[c(1, 4, 5, 6, 7, 10, 11, 12)],
  ssps = list_ssps()[2],
  obs_periods = list_obs_periods()[1],
  gcm_periods = list_gcm_periods(),
  max_run = 10,
  cache = TRUE,
  ...
)
}
\arguments{
\item{xyz}{a \code{terra::SpatRaster} with a single layer containing elevation values in metres, or a \code{data.frame} with the following columns
"long", "lat", "elev", and a unique "id". Any extra columns will be ignored and not output.}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{obs_periods}{character. Which historical period for observational climate
data, averaged over this period. Options are \code{\link[=list_obs_periods]{list_obs_periods()}}.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}.}

\item{max_run}{integer. Maximum number of model runs to include, not including the ensemble mean.
Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{...}{other arguments passed to \code{\link[=downscale_core]{downscale_core()}}. Namely: \code{return_refperiod},
\code{vars}, \code{out_spatial} and \code{plot}}
}
\description{
This is a wrapper function that will create the climate diagram input and
then will call the function to generate the climate diagram.
}
\examples{
\dontrun{
in_xyz <- data.frame(lon = -127.7052, lat = 55.3557, elev = 291, id = 1)
plot_climate_diagram(in_xyz)

plot_climate_diagram(in_xyz, 
  gcms = "CanESM5",
  ssp = "ssp245",
  obs_period = "2001_2020",
  gcm_period = "2021_2040",
  max_run = 1,
  cache = TRUE
  )
}
}
