% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_stripes.R
\name{plot_climate_stripes}
\alias{plot_climate_stripes}
\title{Plot climate stripes}
\usage{
plot_climate_stripes(
  X,
  mode = c("warming_stripes", "labelled_stripes", "bars", "bars_with_scale")
)
}
\arguments{
\item{mode}{A character string. Either \code{warming_stripes}, \code{labelled_stripes}, \code{bars} or
\code{bars_with_scale}.}

\item{x}{A \code{data.table} object that contains the change in temperature relative to the average
and the corresponding years.}
}
\value{
plotly object, Draws a plot in the active graphics device.
}
\description{
Plot temperature change relative to average over time
}
\examples{
if (FALSE) {
}

}
