% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries.R
\name{plot_timeSeries}
\alias{plot_timeSeries}
\title{Time series plots of climate change}
\usage{
plot_timeSeries(
  xyz,
  variable1 = "Tmin_sm",
  variable2 = NULL,
  showObserved = T,
  historic_ts_dataset = "climate_na",
  gcms.ts = list_gcm()[c(1, 4, 5, 7, 10, 11, 12)],
  ssps = list_ssp()[1:3],
  showrange = T,
  yfit = T,
  cex = 1,
  mar = c(3, 3, 0.1, 4),
  showmean = T,
  compile = T,
  simplify = T,
  refline = T,
  refline.obs = T,
  label.endyear = F,
  endlabel = "change",
  yearmarkers = T,
  yearlines = F,
  legend_pos = "topleft"
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{variable1}{character. A climate variable. options are \code{list_variables()}.}

\item{variable2}{character. A second climate variable to plot in combination with \code{variable1}.
options are \code{list_variables()}.}

\item{showObserved}{logical. Plot a time series of observed climate.}

\item{historic_ts_dataset}{character. The observational time series dataset specified in the
\code{climr::plot_timeSeries_input()} call used to create the \code{X} input table. Defaults to \code{"climate_na"}.
This parameter is only used for specifying the correct dataset in the plot legend.}

\item{showrange}{logical. Plot a shaded region indicating the minimum and maximum of the
selected ensemble of GCM simulations for each selected scenario.}

\item{yfit}{logical. Set the range of the y axis to the range of the visible data. If \code{FALSE}
the y axis is the range of all values of \code{variable1} (and \code{variable2} if applicable) in the
input table defined by \code{X}.}

\item{showmean}{logical. Plot the ensemble mean time series. Multi-model ensemble means are
calculated from the mean of simulations for each model.}

\item{compile}{logical. Compile multiple global climate models into a multi-model ensemble.
If \code{FALSE} the single-model ensembles are plotted individually.}

\item{simplify}{logical. Simplify the ensemble range and mean using a smoothing spline.}

\item{refline}{logical. Plot the 1961-1990 reference period mean for the selected variable
and extend this line to the year 2100 as a visual reference.}

\item{refline.obs}{logical. Plot the 1961-1990 reference period mean for the observational data.
This should be the same as the reference line for the GCM time series.}

\item{label.endyear}{logical. Add a label of the final year of the observational time series.}

\item{endlabel}{character. Add a label to the end of each simulated time series. Options
are "change", to indicate the change in year 2100 relative to the 1961-1990 baseline, or "model"
to indicate the global climate model.}

\item{yearmarkers}{logical. Add white points to the observational time series as a visual aid.}

\item{yearlines}{logical. Add vertical lines on every fifth year as a visual reference}

\item{legend_pos}{character. Position of the legend. Viable options are \verb{c("bottomright", "bottomleft", "topleft", and "topright")}.}

\item{X.}{A \code{data.table} produced using the function \code{climr::plot_timeSeries_input()}. This
table can include more models, scenarios, and variables than are used in individual calls to
\code{plot_timeSeries}.}
}
\value{
NULL. Draws a plot in the active graphics device.
}
\description{
Time series plots of 20th and 21st century climate change for user-selected locations and climate variables.
The purposes of the plot are to:
\enumerate{
\item view differences in interannual variability and climate change trends among global climate models (GCMs),
\item view the differences between multiple simulations of each model,
\item compare simulated and observed climate change from 1901 to present, and
\item compare time series of two different variables.
}
All global climate model anomalies are bias-corrected to the 1961-1990 reference period normals.
}
\details{
The input table \code{X} provides climate data for a single location or the average of multiple locations.
The purpose of conducting the generation of the input table in a separate function is to allow users
to make multiple calls to \code{plot_timeSeries} without needing to generate the inputs each time.

Some combinations of \code{variable2} and \code{variable2} are not compatible or meaningful.
Examples of meaningful combinations are winter vs summer values of the same climate variable or minimum vs.
maximum temperatures.
}
\examples{
# data frame of arbitrary points
my_points <- data.frame(lon = c(-127.7300,-127.7500), lat = c(55.34114, 55.25), elev = c(711, 500), id = 1:2)

# generate the input data
my_data <- plot_timeSeries_input(my_points)

# use the input to create a plot
plot_timeSeries(my_data)

# compare mean daily minimum and maximum temperatures
plot_timeSeries(my_data, variable1 = "Tmin_sm", variable2 = "Tmax_sm")

# compare summer and winter temperatures
plot_timeSeries(my_data, variable1 = "Tmax_sm", variable2 = "Tmax_wt")

# compare global climate models
plot_timeSeries(my_data, variable1 = "Tmin_sm", gcms.ts = list_gcm()[c(13,7)], ssp = list_ssp()[2], compile = F, simplify = F, endlabel = "model", mar=c(3,3,0.1,6), showObserved = F)

# export plot to a temporary directory
figDir <- tempdir()
png(
  filename = file.path(figDir, "plot_test.png"), type = "cairo", units = "in",
  width = 6, height = 5, pointsize = 10, res = 300
)
plot_timeSeries(my_points, variable1 = "Tmin_sm")
dev.off()

}
