% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries_input_preprocess.R
\name{plot_timeSeries_input_preprocess}
\alias{plot_timeSeries_input_preprocess}
\title{Time series plots of climate change}
\usage{
plot_timeSeries_input_preprocess(
  X,
  var1 = "Tmin_sm",
  var2 = NULL,
  showObserved = TRUE,
  obs_ts_dataset = "mswx.blend",
  gcms = list_gcms()[c(1, 4, 5, 6, 7, 10, 11, 12)],
  ssps = list_ssps()[1:3],
  showrange = TRUE,
  yfit = TRUE,
  cex = 1,
  mar = c(3, 3, 0.1, 4),
  showmean = TRUE,
  compile = TRUE,
  simplify = TRUE,
  refline = FALSE,
  refline.obs = TRUE
)
}
\arguments{
\item{X}{A \code{data.table}.}

\item{var1}{character. A climate var. options are \code{\link[=list_vars]{list_vars()}}.}

\item{var2}{character. A second climate var to plot in combination with \code{var1}.
options are \code{\link[=list_vars]{list_vars()}}.}

\item{showObserved}{logical. Plot a time series of observed climate.}

\item{obs_ts_dataset}{character. The dataset(s) to use for observational time series data. Options
are \code{[}list_obs_ts_datasets()`].}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{showrange}{logical. Plot a shaded region indicating the minimum and maximum of the
selected ensemble of GCM simulations for each selected scenario.}

\item{yfit}{logical. Set the range of the y axis to the range of the visible data. If \code{FALSE}
the y axis is the range of all values of \code{var1} (and \code{var2} if applicable) in the
input table defined by \code{X}.}

\item{cex}{Numeric. The magnification factor for text size. Default is 1.}

\item{mar}{A numerical vector of length 4, giving the margin sizes in number of lines of text:
c(bottom, left, top, right). The default is c(3,3,0.1,4).}

\item{showmean}{logical. Plot the ensemble mean time series. Multi-model ensemble means are
calculated from the mean of simulations for each model.}

\item{compile}{logical. Compile multiple global climate models into a multi-model ensemble.
If \code{FALSE} the single-model ensembles are plotted individually.}

\item{simplify}{logical. Simplify the ensemble range and mean using a smoothing spline.}

\item{refline}{logical. Plot the 1961-1990 reference period mean for the selected var
and extend this line to the year 2100 as a visual reference.}

\item{refline.obs}{logical. Plot the 1961-1990 reference period mean for the observational data.
This should be the same as the reference line for the GCM time series..}
}
\value{
NULL.
}
\description{
Time series plots of 20th and 21st century climate change for user-selected
locations and climate variables.

The purposes of the plot are to:
\enumerate{
\item view differences in interannual variability and climate change trends among global
climate models (GCMs),
\item view the differences between multiple simulations of each model,
\item compare simulated and observed climate change from 1901 to present, and
\item compare time series of two different variables.
}
All global climate model anomalies are bias-corrected to the 1961-1990 reference period normals.
}
\details{
The input table \code{X} provides climate data for a single location or the average of multiple
locations. The purpose of conducting the generation of the input table in a separate function is
to allow users to make multiple calls to \code{\link[=plot_timeSeries]{plot_timeSeries()}} without needing to generate the
inputs each time.

Some combinations of \code{var1} and \code{var2} are not compatible or meaningful.
Examples of meaningful combinations are winter vs summer values of the same climate var
or minimum vs. maximum temperatures.

Downloads of GCM time series take some time. The \code{plot_timeSeries_input()} function can take
~5 minutes to run for the first time it is called for a location. Once the time series are
cached, they don't need to be downloaded again.
}
\keyword{internal}
