% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WalterLieth_input.R
\name{plot_WalterLieth_input}
\alias{plot_WalterLieth_input}
\title{Input data for the Walter-Lieth climate diagram}
\usage{
plot_WalterLieth_input(
  xyz,
  obs_period = NULL,
  gcms = NULL,
  ssps = NULL,
  gcm_periods = NULL,
  db_option = "auto"
)
}
\arguments{
\item{xyz}{a \code{terra::SpatRaster} with a single layer containing elevation values in metres, or a \code{data.frame} with the following columns
"long", "lat", "elev", and a unique "id". Any extra columns will be ignored and not output.}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}.}

\item{db_option}{character. One of \code{auto}, \code{database}, or \code{local}. Default \code{auto}.}

\item{vars}{character. A vector of climate variables to compute. Supported variables
can be obtained with \code{\link[=list_vars]{list_vars()}}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}
}
\value{
\code{data.table} of average downscaled climate variables for all locations.
}
\description{
Input data for the \code{\link[=plot_WalterLieth]{plot_WalterLieth()}} function. Since these inputs are time-consuming to
generate, the purpose of conducting the generation of the input table in a separate function is
to allow users to make multiple calls to \code{\link[=plot_WalterLieth]{plot_WalterLieth()}} (e.g., for comparing different
time periods) without needing to generate the inputs each time.
}
\details{
This function generates standardized inputs for one or multiple locations at any spatial scale.
}
\examples{
if (FALSE) {
  # data frame of arbitrary points
  my_points <- data.frame(
    lon = c(-127.7300, -127.7500),
    lat = c(55.34114, 55.25),
    elev = c(711, 500),
    id = 1:2
  )

  # generate the input data
  my_data <- plot_WalterLieth_input(my_points, db_option = "local")

  # use the input to create a plot
  plot_WalterLieth(my_data)
}
#'
}
