% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{point_grid}
\alias{point_grid}
\title{Example grid of point locations for sampling}
\format{
\code{PackedSpatVector}
}
\source{
derived from \code{dem_vancouver} reprojected at 0.01 degrees.
}
\usage{
point_grid
}
\description{
Shapefile with point locations spaced at regular intervals, in
North Vancouver, BC, Canada.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{point_grid <- terra::unwrap(get(data(point_grid)))}
}
\keyword{datasets}
