% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.climr_bayes.R
\name{print.summary.climr_bayes}
\alias{print.summary.climr_bayes}
\title{Print method for objects of class \code{"summary.climr_bayes"}}
\usage{
\method{print}{summary.climr_bayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.climr_bayes"}, a result of a call
to \code{\link{summary.climr_bayes}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for an object \code{x} of class
\code{"summary.climr_bayes"}.
}
\details{
Prints the call and the numeric matrix \code{x$matrix} returned from
  \code{\link{summary.climr_bayes}}.
}
\section{Examples}{
 See the examples in \code{\link{fit_bayes}}.
}

\seealso{
\code{\link{fit_bayes}} for estimations from Bayesian inference for a
  two-way random-effects ANOVA model using Markov chain Monte Carlo (MCMC)
  techniques,.

\code{\link{print.climr_bayes}} \code{print} method for class
  \code{"climr_bayes"}.

\code{\link{summary.climr_bayes}} \code{summary} method for class
  \code{"climr_bayes"}.
}
