% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.climr_reml.R
\name{print.summary.climr_reml}
\alias{print.summary.climr_reml}
\title{Print method for objects of class \code{"summary.climr_reml"}}
\usage{
\method{print}{summary.climr_reml}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.climr_reml"}, a result of a call to
\code{\link{summary.climr_reml}}.}

\item{...}{Additional arguments passed on to \code{\link{print.default}}.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for an object \code{x} of class \code{"summary.climr_reml"}.
}
\details{
Prints the call and the numeric matrix \code{x$matrix} returned from
  \code{\link{summary.climr_reml}}.
}
\section{Examples}{
 See the examples in \code{\link{fit_reml}}.
}

\seealso{
\code{\link{fit_reml}} for estimation of a fitted ANOVA model using
  the restricted maximum likelihood(REML).

\code{\link{print.climr_reml}} \code{print} method for class
  \code{"climr_reml"}.

\code{\link{summary.climr_reml}}: \code{summary} method for class
  \code{"climr_reml"}.
}
