% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{PRISM}
\alias{PRISM}
\alias{extract_PRISM}
\alias{split_PRISM_fname}
\title{Functions for PRISM manipulation}
\usage{
extract_PRISM(fl, coords, by)

split_PRISM_fname(fl_list, var = 2L)
}
\description{
NA
}
\author{
Brandon McNellis
}
