% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries_input_preprocess.R
\name{process_ensemble}
\alias{process_ensemble}
\title{Function for processing time series data for gcms or compiled ensemble}
\usage{
process_ensemble(
  dt,
  x,
  var,
  scenarios.selected,
  scenarios,
  showrange = TRUE,
  simplify = TRUE,
  gcm = NULL,
  refline = FALSE,
  showmean = TRUE,
  element,
  compile = TRUE,
  var2 = NULL,
  element1,
  element2,
  yeartime.names,
  yeartimes,
  yeartime
)
}
\arguments{
\item{dt}{empty data table}

\item{x}{climate data}

\item{var}{TODO}

\item{scenarios.selected}{TODO}

\item{scenarios}{TODO}

\item{showrange}{logical. Plot a shaded region indicating the minimum and maximum of the
selected ensemble of GCM simulations for each selected scenario.}

\item{simplify}{logical. Simplify the ensemble range and mean using a smoothing spline.}

\item{gcm}{TODO}

\item{refline}{logical. Plot the 1961-1990 reference period mean for the selected var
and extend this line to the year 2100 as a visual reference.}

\item{showmean}{logical. Plot the ensemble mean time series. Multi-model ensemble means are
calculated from the mean of simulations for each model.}

\item{element}{TODO}

\item{compile}{logical. Compile multiple global climate models into a multi-model ensemble.
If \code{FALSE} the single-model ensembles are plotted individually.}

\item{var2}{character. A second climate var to plot in combination with \code{var1}.
options are \code{\link[=list_vars]{list_vars()}}.}

\item{element1}{TODO}

\item{element2}{TODO}

\item{yeartime.names}{TODO}

\item{yeartimes}{TODO}

\item{yeartime}{TODO}
}
\description{
Function for processing time series data for gcms or compiled ensemble
}
