% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{process_one_climaterast}
\alias{process_one_climaterast}
\title{TODO: fill documentation here}
\usage{
process_one_climaterast(
  climaterast,
  res,
  xyz,
  timeseries = FALSE,
  type = c("gcm", "gcm_hist", "historic")
)
}
\arguments{
\item{climaterast}{}

\item{res}{}

\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{timeseries}{}

\item{type}{}
}
\value{
a \code{data.table}
}
\description{
TODO: fill documentation here
}
