% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{process_one_gcm}
\alias{process_one_gcm}
\title{TODO: fill documentation here}
\usage{
process_one_gcm(gcm_, res, xyzID, timeseries)
}
\arguments{
\item{gcm_}{TODO}

\item{res}{TODO}

\item{xyzID}{TODO}

\item{timeseries}{TODO}
}
\value{
\code{data.table}
}
\description{
TODO: fill documentation here
}
