% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{process_one_gcm2}
\alias{process_one_gcm2}
\title{TODO: add documentation here}
\usage{
process_one_gcm2(
  gcm_nm,
  ssp,
  bbox,
  period,
  max_run,
  dbnames = dbnames,
  dbCon,
  cache
)
}
\arguments{
\item{gcm_nm}{character. Name of GCM to use.}

\item{ssp}{A character vector. Label of the shared socioeconomic pathway(s) to use.
See available options with \code{list_ssp()}. Defaults to \code{list_ssp()}.}

\item{bbox}{Numeric vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{period}{A character vector. Label of the period to use.
Can be obtained from \code{list_gcm_period()}. Default to \code{list_gcm_period()}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{dbnames}{\code{data.frame} with the list of available GCMs and their
corresponding names in the PostGIS data base.  See climr:::dbnames}

\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{cache}{Logical specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
Spat Raster
}
\description{
TODO: add documentation here
}
