% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{process_one_gcm3}
\alias{process_one_gcm3}
\title{Process one historic time series at a time}
\usage{
process_one_gcm3(
  gcm_nm,
  years,
  dbCon,
  bbox,
  max_run,
  dbnames = dbnames_hist,
  cache
)
}
\arguments{
\item{gcm_nm}{character. Name of GCM to use.}

\item{years}{numeric. Vector of desired years. Must be in \code{1851:2015}.
Can be obtained from \code{list_gcm_period()}. Default to \code{list_gcm_period()}.}

\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{numeric. Vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{dbnames}{\code{data.frame} with the list of available GCMs (historical projections)
and their corresponding names in the PostGIS data base. See climr:::dbnames_hist}

\item{cache}{logical. Specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
\code{SpatRaster}
}
\description{
Process one historic time series at a time
}
