% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{process_one_gcm4}
\alias{process_one_gcm4}
\title{Process one GCM time series at a time}
\usage{
process_one_gcm4(
  gcm_nm,
  ssp,
  period,
  max_run,
  dbnames = dbnames_ts,
  bbox,
  dbCon,
  cache
)
}
\arguments{
\item{gcm_nm}{character. Name of GCM to use.}

\item{ssp}{character. Vector of labels of the shared socioeconomic pathways to use.
See available options with \code{list_ssp()}. Defaults to \code{list_ssp()}.}

\item{period}{character. Vector of labels of the periods to use.
Can be obtained from \code{list_gcm_period()}. Default to \code{list_gcm_period()}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{dbnames}{\code{data.frame} with the list of available GCMs (time series projections)
and their corresponding names in the PostGIS data base. See climr:::dbnames_ts}

\item{bbox}{numeric. Vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{cache}{logical. Specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
a \code{SpatRaster}
}
\description{
Process one GCM time series at a time
}
