% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{process_one_gcm_hist}
\alias{process_one_gcm_hist}
\title{TODO: fill documentation here}
\usage{
process_one_gcm_hist(gcm_, res, xyzID)
}
\arguments{
\item{gcm_}{}

\item{res}{}

\item{xyzID}{}
}
\value{
\code{data.table}
}
\description{
TODO: fill documentation here
}
