% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{process_one_historic}
\alias{process_one_historic}
\title{TODO: fill documentation here}
\usage{
process_one_historic(historic_, res, xyzID, timeseries)
}
\arguments{
\item{historic_}{}

\item{res}{}

\item{xyzID}{}

\item{timeseries}{}
}
\value{
\code{data.table}
}
\description{
TODO: fill documentation here
}
