% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector.R
\name{scenarioMatrix}
\alias{scenarioMatrix}
\title{scenarioMatrix}
\usage{
scenarioMatrix(scenarios, ecs_0)
}
\arguments{
\item{scenarios}{scenarios data frame}

\item{ecs_0}{climate sensitivity}
}
\value{
data frame : year, variable, global value, units, global value excluding Ireland, national contributin, national scenario, global scenario, ecs
}
\description{
combined national and 4 global scenarios from AR4
}
\examples{
scenarioMatrix(scenarios,2.9)

}
