% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_ipcc.R
\docType{data}
\name{scenarios_ipcc}
\alias{scenarios_ipcc}
\title{Future emissions scenarios for Ireland}
\format{
A data frame with 225 rows and 6 variables:
\describe{
\item{scenario}{national emissions scenario}
\item{gas}{climate forcer}
\item{year}{year}
\item{factor}{emissions as a fraction of 2018 value}
\item{value}{actual emissions in IPCC units}
\item{units}{unit}
}
}
\source{

}
\usage{
scenarios_ipcc
}
\description{
Nine scenarios A-I with emissions expressed as fraction of 2018 for years 2025,2030,2035,2050,2100
Emissions also expressed in absolute IPCC units
}
\keyword{datasets}
