% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2_irl.R
\docType{data}
\name{so2_irl}
\alias{so2_irl}
\title{Estimated SO2 Emissions from Ireland 1745-2018}
\format{
A data frame with 274 rows and 2 variables:
\describe{
\item{year}{year}
\item{Gg S}{Gg Sulphur}
}
}
\source{
1990-2018 \url{https://stats.oecd.org/Index.aspx?DataSetCode=AIR_EMISSIONS} 1850-1989
\url{https://sedac.ciesin.columbia.edu/data/set/haso2-anthro-sulfur-dioxide-emissions-1850-2005-v2-86/data-download}
pre-1850 rescaled global estimate
}
\usage{
so2_irl
}
\description{
Units Gg S equivalent to 0.5 kt SO2
}
\keyword{datasets}
