% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_downscale_ouput.R
\name{summarize_downscale_output}
\alias{summarize_downscale_output}
\title{Summarize downscale for plot inputs}
\usage{
summarize_downscale_output(
  xyz = NULL,
  gcms = list_gcms()[c(1, 4, 5, 6, 7, 10, 11, 12)],
  ssps = list_ssps()[2],
  max_run = 10L,
  obs_years = list_obs_years(),
  gcm_hist_years = list_gcm_hist_years(),
  gcm_ssp_years = list_gcm_ssp_years(),
  use_downscale_db = FALSE,
  cache = TRUE,
  vars = NULL,
  downscale_results = NULL,
  ...
)
}
\arguments{
\item{xyz}{a \code{terra::SpatRaster} with a single layer containing elevation values in metres, or a \code{data.frame} with the following columns
"long", "lat", "elev", and a unique "id". Any extra columns will be ignored and not output.}

\item{gcms}{character. Vector of global climate model names. Options are \code{\link[=list_gcms]{list_gcms()}}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{max_run}{integer. Maximum number of model runs to include, not including the ensemble mean.
Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{obs_years}{integer. Vector of years to obtain individual years or time
series of observational climate data. See \code{\link[=list_obs_years]{list_obs_years()}} for available years.}

\item{gcm_hist_years}{character.  Timeseries years for GCM simulations of the
historical scenario. See \code{\link[=list_gcm_hist_years]{list_gcm_hist_years()}} for available years.}

\item{gcm_ssp_years}{character. Timeseries years for GCM simulations of future
scenarios specified by \code{ssps}. See \code{\link[=list_gcm_ssp_years]{list_gcm_ssp_years()}} for available years.}

\item{use_downscale_db}{Should the function \code{downscale_db} be used instead of \code{downscale}}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{vars}{character vector of the vars that needs to be summarize}

\item{downscale_results}{optional, if you want to pass precomputed downscale results}

\item{...}{Additional arguments passed to \code{downscale} or \code{downscale_db}}
}
\value{
A data.table containing averaged climate variables by period,
including monthly temperature (average, min, max) and precipitation
values (Tave, Tmin, Tmax, PPT) for each of the 12 months
}
\description{
This function will summarize downscale to prepare input for plots
}
\examples{
\dontrun{
in_xyz <- data.frame(lon = -127.7052, lat = 55.3557, elev = 291, id = 1)
summarize_downscale_output(in_xyz)
summarize_downscale_output(in_xyz, max_run = 1L, 
  gcm_hist_years = 1950:2015, 
  gcm_ssp_years = 2015:2040)
}
}
\keyword{internal}
