% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.climr_bayes.R
\name{summary.climr_bayes}
\alias{summary.climr_bayes}
\title{Summary method for an \code{"climr_bayes"} object}
\usage{
\method{summary}{climr_bayes}(x, digits = max(3, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{"climr_bayes"}, a result of a call to
\code{\link{fit_bayes}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{summary}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
Returns an object (a list) of class \code{"summary.climr_bayes"}
  containing the list element \code{x$call} and a numeric matrix
  \code{matrix} giving, for all five parameters, their estimators, the
  corresponding estimated standard errors (Std. Error) and 95\% confidence
  intervals i.e. the values subtracted from the raw estimate.
}
\description{
\code{summary} method for class \code{"climr_bayes"}
}
\section{Examples}{
 See the examples in \code{\link{fit_bayes}}.
}

\seealso{
\code{\link{fit_bayes}} for estimations from Bayesian inference for
  a two-way random-effects ANOVA model using Markov chain Monte Carlo (MCMC)
  techniques,.

\code{\link{print.climr_bayes}} \code{print} method for class
  \code{"climr_bayes"}.
}
