% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.climr_reml.R
\name{summary.climr_reml}
\alias{summary.climr_reml}
\title{Summary method for an \code{"climr_reml"} object}
\usage{
\method{summary}{climr_reml}(x, digits = max(3, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{"climr_reml"}, a result of a call to
\code{\link{fit_reml}}.}

\item{digits}{An integer. Used for number formatting with
\code{\link[base]{summary}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
Returns an object (a list) of class \code{"summary.climr_reml"}
  containing the list element \code{x$call} and a numeric matrix
  \code{matrix} giving, for all five parameters, their estimators and the
  corresponding estimated standard errors (Std. Error) 95\% confidence
  intervals i.e. the values subtracted from the raw estimate.
}
\description{
\code{summary} method for class \code{"climr_reml"}
}
\section{Examples}{
 See the examples in \code{\link{fit_reml}}.
}

\seealso{
\code{\link{fit_reml}} for estimation of a fitted ANOVA model using
  the restricted maximum likelihood(REML).

\code{\link{print.climr_reml}} \code{print} method for class
  \code{"climr_reml"}.
}
