% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cmip5.R
\name{summary.cmip5}
\alias{summary.cmip5}
\title{Descriptive summary for CMIP5 data}
\usage{
\method{summary}{cmip5}(data, ...)
}
\arguments{
\item{data}{An object of class \code{"cmip5"}, a result of a call to
\code{\link{nnt}}.}

\item{...}{Further arguments to be passed to \code{\link[base]{summary}}}
}
\value{
return the table \code{k} that contains the overall organised table
  for numbers of run for each combination of GCMs and RCPs
}
\description{
\code{summary} method for an object of class \code{"cmip5"},following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Richard
(2014)}
}
\details{
This function is only applicable for the CMIP5 data containing the
  index, GCM and RCP; Users could obtain an organized table to summary the
  number of the runs for each combinations of GCMs and RCPs(i.e. RCP2.6,
  RCP4.5, RCP6.0 and RCP8.5).
}
\examples{
summary(cmip5_temp1)

}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying Sources of
  Uncertainty in Projections of Future Climate. \emph{Journal of Climate},
  \strong{27}, 8793-8808. \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip5_temp1}} and \code{\link[climr]{cmip5_temp2}}
 for CMIP5 data.
}
