% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cmip6.R
\name{summary.cmip6}
\alias{summary.cmip6}
\title{Descriptive plot for CMIP6 data}
\usage{
\method{summary}{cmip6}(x, region = 0, period = 1, ...)
}
\arguments{
\item{region}{Region of the world: 0 for global, 1-22 for regions.(default 0)}

\item{period}{Time horizon: 1 for mid 21st century (2020-2049); 2 for late
21st century (2070-2099).}

\item{...}{Further arguments to be passed to \code{\link[base]{summary}}}

\item{data}{An object of class \code{"cmip6"}, a result of a call to
\code{\link{nnt}}.}
}
\value{
return the table \code{k} that contains the overall organised table
  for numbers of run for each combination of GCMs and SSPs in a specific region
  and specific time horizon
}
\description{
\code{summary} method for an object of class \code{"cmip6"},following
\href{http://dx.doi.org/10.1175/JCLI-D-14-00265.1}{Northrop and Richard
(2014)}
}
\details{
This function is only applicable for the CMIP6 data containing the
  index, GCM and SSP; Users could obtain an organized table to summary the
  number of the runs for each combinations of GCMs and SSPs in a specific region
  and specific time horizon.
}
\examples{
summary(cmip6_pr)

}
\references{
Northrop, P.J. and R.E. Chandler (2014). Quantifying Sources of
  Uncertainty in Projections of Future Climate. \emph{Journal of Climate},
  \strong{27}, 8793-8808. \url{https://doi.org/10.1175/JCLI-D-14-00265.1}
}
\seealso{
\code{\link[climr]{cmip6_pr}} and \code{\link[climr]{cmip6_tas}} for
  CMIP6 data.
}
