% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{threaded_downscale_}
\alias{threaded_downscale_}
\title{Wrapper function for \code{downscale_} for parallelising}
\usage{
threaded_downscale_(
  xyz,
  normal,
  gcm,
  gcm_ts,
  gcm_hist,
  historic,
  historic_ts,
  ...
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{normal}{\code{SpatRaster}. Outputs from \code{\link[=normal_input]{normal_input()}}. The high-resolution climate maps for use as the downscaling baseline.}

\item{gcm}{\code{list} of \code{SpatRasters}. Outputs from \code{\link[=gcm_input]{gcm_input()}}. Global climate model data for 20-year normal periods to be downscaled. Default to \code{NULL}.}

\item{gcm_ts}{\code{list} of \code{SpatRasters}. Outputs from \code{\link[=gcm_ts_input]{gcm_ts_input()}}. Global climate model time series for ssp-rcp scenarios to be downscaled. Default to \code{NULL}.}

\item{gcm_hist}{\code{list} of \code{SpatRasters}. Outputs from \code{\link[=gcm_hist_input]{gcm_hist_input()}}. Global climate model time series for historical scenario to be downscaled. Default to \code{NULL}.}

\item{historic}{\code{list} of \code{SpatRasters}. Outputs from \code{\link[=historic_input]{historic_input()}}. Observed climate data for 20-year normal periods to be downscaled. Default to \code{NULL}.}

\item{historic_ts}{\code{list} of \code{SpatRasters}. Outputs from \code{\link[=historic_input_ts]{historic_input_ts()}}. Observed climate time series to be downscaled. Default to \code{NULL}.}

\item{...}{further arguments passed to \code{downscale_}}
}
\value{
A \code{data.table}
}
\description{
Wrapper function for \code{downscale_} for parallelising
}
