% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tif_folder_gen}
\alias{tif_folder_gen}
\title{Process a tif folder of TminXX.tif, TmaxXX.tif, PPTXX.tif
to create a set of climate variables tifs}
\usage{
tif_folder_gen(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{A folder path.}
}
\description{
Process a tif folder of TminXX.tif, TmaxXX.tif, PPTXX.tif
to create a set of climate variables tifs
}
