% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{vancouver}
\alias{vancouver}
\title{Example vector data}
\format{
a \code{SpatVector}
}
\source{
todo add/describe source
}
\usage{
vancouver
}
\description{
Shapefile of coast of Vancouver, BC, Canada.
}
\keyword{datasets}
