% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{vancouver_points}
\alias{vancouver_points}
\title{Example point locations}
\format{
\code{PackedSpatVector}
}
\source{
derived from \code{dem_vancouver}
}
\usage{
vancouver_points
}
\description{
Shapefile with point locations North Vancouver, BC, Canada.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{vancouver_points <- terra::unwrap(get(data(vancouver_points)))}
}
\keyword{datasets}
