% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{vancouver_poly}
\alias{vancouver_poly}
\title{Example polygon locations}
\format{
\code{PackedSpatVector}
}
\source{
derived from \code{dem_vancouver}
}
\usage{
vancouver_poly
}
\description{
Shapefile of North Vancouver in polygon format, BC, Canada.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{vancouver_poly <- terra::unwrap(get(data(vancouver_poly)))}
}
\keyword{datasets}
