% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-package.R
\docType{data}
\name{variables}
\alias{variables}
\title{Climate variables definition}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 234 rows and 5 columns.
}
\usage{
variables
}
\description{
Climate variables definition
}
\keyword{datasets}
