% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{weather_stations}
\alias{weather_stations}
\title{BC Weather Station Data - PCDS}
\format{
a \code{PackedSpatVector}
}
\source{
\url{https://services.pacificclimate.org/met-data-portal-pcds/app/}
}
\usage{
weather_stations
}
\description{
A points shapefile of weather stations in BC from the
Pacific Climate Impacts Consortium. Stations without elevation
data were excluded.
}
\details{
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{weather_stations <- terra::unwrap(get(data(weather_stations)))}
}
\keyword{datasets}
