% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{xyzDT}
\alias{xyzDT}
\title{Example table of point locations}
\format{
a \code{data.table}
}
\source{
\url{https://services.pacificclimate.org/met-data-portal-pcds/app/}
}
\usage{
xyzDT
}
\description{
A \code{data.table} of point locations with elevation.
It is a simplified version of the \code{weather_stations} dataset.
}
\keyword{datasets}
