% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_clim_vars.R
\name{append_clim_vars}
\alias{append_clim_vars}
\title{Add extra climate variables to a \code{data.table}}
\usage{
append_clim_vars(dt, vars)
}
\arguments{
\item{dt}{A \code{data.table} with TminXX, TmaxXX, PPTXX for XX in 01 to 12.}

\item{vars}{A character vector of climate variables to compute.}
}
\value{
a \code{data.table}
}
\description{
Add extra climate variables to a \code{data.table}
}
