% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_clim_vars.R
\name{append_clim_vars}
\alias{append_clim_vars}
\title{Append Additional Climate Variables
Add extra climate variables to a SpatRaster or \code{data.table}}
\usage{
append_clim_vars(dt, vars)
}
\arguments{
\item{dt}{object on which to add variables. Eith a SpatRaster or a data.table}

\item{vars}{variables to add}
}
\description{
Append Additional Climate Variables
Add extra climate variables to a SpatRaster or \code{data.table}
}
\keyword{internal}
