% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_delete}
\alias{cache_delete}
\alias{cache_exists}
\alias{cache_list}
\alias{cache_path}
\title{Delete package local cache}
\usage{
cache_delete()

cache_exists()

cache_list()

cache_path()
}
\value{
The full path of the package local cache.
}
\description{
Delete package local cache

Check if package local cache exists

List package local cache files

Return package local cache path
}
\details{
By default, it uses \code{tools::R_user_dir}. The cache location can be
set using the \code{climRpnw.cache.path} option with \code{options("climRpnw.cache.path" = "your_path")}.
}
