% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DD.R
\name{calc_DD_above_18}
\alias{calc_DD_above_18}
\title{Calculate Degree-Day Above 18 (DD>18)}
\usage{
calc_DD_above_18(m, tm, region)
}
\arguments{
\item{m}{month of the year}

\item{tm}{monthly mean temperature for the \code{m} month}

\item{region}{one of either "All", "South west", "The rest"}
}
\value{
Degree-Days Above 18
}
\description{
Calculate Degree-Day Above 18 (DD>18)
}
\examples{
\dontrun{
climr:::calc_DD_above_18(2, -14, "All")
}
}
