% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EMT_EXT.R
\name{calc_EMT}
\alias{calc_EMT}
\title{Calculate Extreme Minimum Temperature (EMT)}
\usage{
calc_EMT(t_min_list, td)
}
\arguments{
\item{t_min_list}{\code{list} of monthly minimum temperature for each month in order.}

\item{td}{numeric. Difference between the mean warmest monthly temperature and
the mean coldest monthly temperature.}
}
\value{
numeric. Extreme Minimum Temperature
}
\description{
Calculate Extreme Minimum Temperature (EMT)
}
