% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_Eref}
\alias{calc_Eref}
\title{Calculate Extreme Minimum Temperature (EMT)}
\usage{
calc_Eref(m, tmin, tmax, latitude)
}
\arguments{
\item{m}{month of the year}

\item{tmin}{monthly mean minimum air temperature}

\item{tmax}{monthly mean maximum air temperature}

\item{latitude}{point latitudes}
}
\value{
Reference evaporation (Eref)
}
\description{
Calculate Extreme Minimum Temperature (EMT)
}
