% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PAS.R
\name{calc_PAS}
\alias{calc_PAS}
\title{Calculate Precipitation As Snow (PAS)}
\usage{
calc_PAS(m, tm, ppt)
}
\arguments{
\item{m}{month of the year}

\item{tm}{min temperature for that month}

\item{ppt}{precipitation mm}
}
\value{
Precipitation As Snow
}
\description{
Calculate Precipitation As Snow (PAS)
}
\examples{
\dontrun{
climr:::calc_PAS(4, 2)
}
}
