% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_S0_I}
\alias{calc_S0_I}
\title{PROGRAM I From Hargreaves 1985}
\usage{
calc_S0_I(d, tm, latitude)
}
\arguments{
\item{d}{numeric. Julian day of the year (January 1 = 1, December 31 = 365).}

\item{tm}{numeric. Monthly mean temperature for the corresponding month.}

\item{latitude}{numeric. Point latitudes.}
}
\value{
numeric. Extraterrestrial radiation estimation in mm/day
}
\description{
PROGRAM I From Hargreaves 1985
}
