% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Eref_CMD.R
\name{calc_S0_II}
\alias{calc_S0_II}
\title{PROGRAM II From Hargreaves 1985}
\usage{
calc_S0_II(m, tm, latitude)
}
\arguments{
\item{m}{integer. Month of the year.}

\item{tm}{numeric. Monthly mean temperature for the \code{m} month.}

\item{latitude}{numeric. Point latitudes.}
}
\value{
numeric. Extraterrestrial radiation estimation in mm/day
}
\description{
PROGRAM II From Hargreaves 1985
}
