\name{climr-package}
\alias{climr-package}
\alias{climr}
\docType{package}
\title{
\packageTitle{climr}
}
\description{
\packageDescription{climr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{climr}
\packageIndices{climr}
}
\author{
\packageAuthor{climr}

Maintainer: \packageMaintainer{climr}
}
\keyword{ package }
