% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{climr_downscale}
\alias{climr_downscale}
\title{Downscale and Calculate climate variables for points of interest}
\usage{
climr_downscale(
  xyz,
  which_normal = c("auto", "BC", "NorAm"),
  historic_period = NULL,
  historic_ts = NULL,
  gcm_models = NULL,
  ssp = c("ssp126", "ssp245", "ssp370", "ssp585"),
  gcm_period = NULL,
  gcm_ts_years = NULL,
  gcm_hist_years = NULL,
  max_run = 0L,
  return_normal = TRUE,
  vars = sort(sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))),
  cache = TRUE
)
}
\arguments{
\item{xyz}{three or four column data.frame: long, lat, elev, (id)}

\item{which_normal}{Select which normal layer to use. Default is "auto", which selects the highest resolution normal for each point}

\item{historic_period}{Which historic period. Default \code{NULL}}

\item{historic_ts}{Historic years requested. Must be in \code{1902:2015}. Default \code{NULL}}

\item{gcm_models}{Vector of GCM names. Options are \code{list_gcm()}. Used for gcm periods, gcm timeseries, and historic timeseries. Default \code{NULL}}

\item{ssp}{A character vector. Label of the shared socioeconomic pathway(s) to use.
See available options with \code{list_ssp()}. Defaults to \code{list_ssp()}.}

\item{gcm_period}{Requested future periods. Options are \code{list_gcm_period()}}

\item{gcm_ts_years}{Requested future timeseries years. Must be in \code{2015:2100}}

\item{gcm_hist_years}{Requested historic modelled years. Must be in \code{1851:2010}}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{return_normal}{Logical. Return downscaled normal period (1961-1990). Default \code{TRUE}}

\item{vars}{Character vector of climate variables. Options are \code{list_vars()}}

\item{cache}{Logical. Cache data locally? Default \code{TRUE}}
}
\value{
\code{data.frame} of downscaled climate variables for each location.
Historic, normal, and future periods are all returned in one table.
}
\description{
Downscale and Calculate climate variables for points of interest
}
\details{
Additional details... TODO.
}
\examples{
{
  ## TODO.
}

}
\author{
Kiri Daust
}
