% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_get.R
\name{content_get}
\alias{content_get}
\alias{content_get_gh}
\alias{content_get_void}
\title{Get content from source}
\usage{
content_get(path, source = c("github"), ...)

content_get_gh(
  path,
  repo = getOption("climRpnw.gh.repo", default = "bcgov/climR-pnw"),
  ref = "data"
)

content_get_void(...)
}
\arguments{
\item{path}{File path relative to repo root without starting \code{/}.}

\item{source}{A character. Name of the content source. Only supports
"github" as of now.}

\item{...}{Extra arguments for \code{content_get} source dependent functions.}

\item{repo}{Repository name prefixed with organisation name and separated by a \code{/}.
Default to option "climRpnw.gh.repo" value if set, or "bcgov/climR-pnw". This is
the source root.}

\item{ref}{The name of the commit/branch/tag. Default NULL means it uses the repository’s default branch.}
}
\value{
A list of elements, each with a \code{url}, a \code{path} relative
to data path and a unique identifier \code{uid}.

A list of metadata about files in a repository directory or the content of a file. See
https://docs.github.com/en/rest/reference/repos#get-repository-content for details.
}
\description{
Get content from source
}
