% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lists.R
\name{list_gcms}
\alias{list_gcms}
\alias{list_ssps}
\alias{list_gcm_periods}
\alias{list_runs_ssp}
\alias{list_runs_historic}
\alias{list_refmaps}
\alias{list_obs_periods}
\alias{list_vars}
\alias{list_obs_years}
\alias{list_gcm_ssp_years}
\alias{list_gcm_hist_years}
\title{List available runs, global climate models (GCMs), time periods and scenarios (SSPs)}
\usage{
list_gcms()

list_ssps()

list_gcm_periods()

list_runs_ssp(gcm, ssp)

list_runs_historic(gcm)

list_refmaps()

list_obs_periods()

list_vars(set = c("All", "Monthly", "Seasonal", "Annual"), only_extra = FALSE)

list_obs_years()

list_gcm_ssp_years()

list_gcm_hist_years()
}
\arguments{
\item{gcm}{Name of GCM}

\item{ssp}{Name of scenario Must be one of the elements in list_ssps().}

\item{set}{character. One of \code{c("All", "Monthly", "Seasonal", "Annual")} or any combination thereof. Defaults to "All".}

\item{only_extra}{logical. Should Tmin, Tmax and PPT be excluded? Defaults to FALSE.}
}
\value{
a character vector.
}
\description{
\code{list_gcms} lists available global climate models.

\code{list_ssps} lists available greenhouse gas concentration scenarios (SSP-RCPs).

\code{list_gcm_periods} lists available 20-year normal periods for GCM simulations.

\code{list_runs_ssp} lists available runs for a given GCM/ssp.

\code{list_runs_historic} lists available runs from the historical simulation (1851-2014) for a specified GCM.

\code{list_refmaps} lists available reference maps of gridded climate normals

\code{list_obs_periods} lists available normal periods for observational climate data

\code{list_vars} lists available climate variables

\code{list_obs_years} lists available years for time series of observational climate data

\code{list_gcm_ssp_years} lists available years for time series of global climate model future simulations

\code{list_gcm_hist_years} lists available years for time series of global climate model historical simulations
}
\details{
The eight-gcm ensemble recommended by Mahony et al. (2022) can be specified as \code{gcms = list_gcm()[c(1,4,5,6,7,10,11,12)]}.

Currently available reference maps of gridded climate normals (\code{list_refmaps()}) are:
\itemize{
\item "refmap_climatena" for Climate NA derived normals
\item "refmap_climr" for a composite of BC PRISM, adjusted US PRISM and
DAYMET (Alberta and Saskatchewan), covering North America
}

A lookup table of variable names, units, types, and other attributes is available by calling \code{data(variables)}.
}
