% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lists.R
\name{list_gcm}
\alias{list_gcm}
\alias{list_ssp}
\alias{list_gcm_period}
\alias{list_run}
\alias{list_normal}
\alias{list_historic}
\alias{list_variables}
\title{List available runs, global circulation models, periods and climate scenarios}
\usage{
list_gcm()

list_ssp()

list_gcm_period()

list_run(dbCon, gcm)

list_normal()

list_historic()

list_variables(only_extra = FALSE)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{gcm}{Character vector to specify requested GCMs}

\item{only_extra}{A boolean. Should Tmin, Tmax and PPT be excluded? Defaults to FALSE.}
}
\value{
a character vector.
}
\description{
\code{list_gcm} lists available global circulation models

\code{list_ssp} lists available shared socioeconomic pathways

\code{list_gcm_period} lists available periods

\code{list_run} lists available runs for a given GCM

\code{list_normal} lists available normals

\code{list_historic} lists available historic periods

\code{list_variables} lists climate variables
}
\details{
Currently available normals (\code{list_normal()}) span North America ("normal_na") or
British Columbia ("normal_bc")
}
