% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{data_connect}
\alias{data_connect}
\alias{connections}
\alias{connections_clear}
\title{List connections in cache}
\usage{
data_connect(profile = .globals[["sessprof"]]$list())

connections()

connections_clear()
}
\arguments{
\item{profile}{Either \code{climr-db-user} or \code{local}.}
}
\description{
List connections in cache

List connections in cache

Clear connections in cache
}
