% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_update}
\alias{data_update}
\title{Update external package data}
\usage{
data_update(
  ...,
  gcm = getOption("climr.gcm.path", default = "inputs_pkg/gcm"),
  historic = getOption("climr.historic.path", default = "inputs_pkg/historic"),
  normal = getOption("climr.normal.path", default = "inputs_pkg/normal"),
  quiet = !interactive()
)
}
\arguments{
\item{...}{Others parameters such as \code{source} or \code{repo} for content getting functions.}

\item{gcm}{A character. Relative path from the source root to global circulation models files folder.
Default to option value "climr.gcm.path" if set, or "inputs_pkg/gcmData".}

\item{normal}{A character. Relative path from the source root to base normal files folder.}

\item{quiet}{A logical. If \code{TRUE}, suppress status messages (if any), and the progress bar.}
}
\description{
Update external package data
}
\details{
This package uses data that are too big to be included with sources.
Instead, data is downloaded, optionally cached, when you need to run functions.
}
