% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{dbGetTiles}
\alias{dbGetTiles}
\title{Download individual raster tiles from PostGIS}
\usage{
dbGetTiles(name, pnts, bands = 1:73)
}
\arguments{
\item{name}{character. Table name in database.}

\item{pnts}{data.table of point locations with columns \code{lon} and \code{lat}}

\item{bands}{numeric. Which raster bands to return. Default 37:73.}
}
\value{
A \code{SpatRaster}
}
\description{
Download individual raster tiles from PostGIS
}
\keyword{internal}
