% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dem_to_table}
\alias{dem_to_table}
\title{Extract raster cell centers (lon/lat) and values into a data.table}
\usage{
dem_to_table(dem)

dem_to_table(dem)
}
\arguments{
\item{dem}{A \code{SpatRaster} (terra) digital elevation model}
}
\value{
A \code{data.table} with columns \code{id}, \code{lon}, \code{lat}, and \code{elev}, where
each row corresponds to a raster cell.

A \code{data.table} with columns \code{id}, \code{lon}, \code{lat}, \code{elev}
}
\description{
This function converts a digital elevation model in \code{SpatRaster} to a \code{data.table} containing:
\itemize{
\item \code{id}   – the raster cell index (in terra’s internal order)
\item \code{lon}  – longitude of the cell center (decimal degrees, EPSG:4326)
\item \code{lat}  – latitude of the cell center (decimal degrees, EPSG:4326)
\item \code{elev} – raster values (i.e., elevation)
}

This function converts a digital elevation model in \code{SpatRaster} format to a
\code{data.table} with fields required for \code{climr::downscale()}:
\itemize{
\item \code{id}   – raster cell index
\item \code{lon}  – longitude (decimal degrees, EPSG:4326)
\item \code{lat}  – latitude (decimal degrees, EPSG:4326)
\item \code{elev} – raster value
}
}
\details{
The raster is not reprojected. Instead, the native-CRS cell center
coordinates are transformed on the fly to lon/lat, preserving grid alignment.

The primary use case for this function is to create a climr input table for downscaling
a digital elevation model in a geographic projection such as Albers Equal Area.
Since this function conserves the cell centre locations, downstream analysis results
can be mapped in the original geographic projection.
}
\examples{
\dontrun{
library(terra)
dem <- rast("dem_BC2kmGrid.tif")
dt <- raster_to_table(dem)
head(dt)
}

\dontrun{
library(terra)
dem <- unwrap(climr::dem_vancouver) # read in a digital elevation model
dt <- raster_to_table(dem) #convert the digital elevation model into data table
clim <- downscale(dt) # use the table as input a climr downscaling query

X <- dem # use the dem as a template raster
X <- raster::setValues(X,NA) # clear the values from the template raster
values(X)[dt$id] <- dt$elev # populate the template raster with a climate variables
terra::plot(X) # plot a map of the climate values
}
}
