% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{dem_vancouver}
\alias{dem_vancouver}
\title{Example DEM}
\format{
a \code{PackedSpatRaster}
}
\source{
todo add/describe source
}
\usage{
dem_vancouver
}
\description{
Digital elevation model of North Vancouver, BC, Canada.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{dem_vancouver <- terra::unwrap(get(data(dem_vancouver)))}
}
\keyword{datasets}
