% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Self-contained change-factor downscaling of observed and simulated climate data}
\usage{
downscale(
  xyz,
  which_refmap = "auto",
  obs_periods = NULL,
  obs_years = NULL,
  obs_ts_dataset = NULL,
  gcms = NULL,
  ssps = NULL,
  gcm_periods = NULL,
  gcm_ssp_years = NULL,
  gcm_hist_years = NULL,
  max_run = 0L,
  cache = TRUE,
  ...
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{which_refmap}{character. Which climatological normals map to use as the
high-resolution reference climate map for downscaling. Default is "auto",
which selects, for each query point, the best available climatological
normals map in declining order of normals_bc, normals_composite, and normals_na.
Other options are one of \code{\link[=list_refmaps]{list_refmaps()}}, which will provide a consistent
reference map for all points.}

\item{obs_periods}{character. Which obs period for observed climate
data, averaged over this period. Options are \code{\link[=list_obs_periods]{list_obs_periods()}}. Default \code{NULL}}

\item{obs_years}{integer. Vector of years to obtain individual years or time
series of observed climate data. Default \code{NULL}. See \code{\link[=list_obs_years]{list_obs_years()}}
for available years.}

\item{obs_ts_dataset}{character. The dataset to use for observed time series data. Options
are \code{"climatena"} for the ClimateNA gridded time series or \code{"cru.gpcc"} for the combined Climatic
Research Unit TS dataset (for Temperature) and Global Precipitation Climatology Centre dataset
(for precipitation). Defaults to \code{NULL}.}

\item{gcms}{character. Vector of global climate model names. Options
are \code{\link[=list_gcms]{list_gcms()}}. Used for gcms periods, gcms timeseries, and obs timeseries.
Defaults to \code{NULL}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}. Defaults to \code{NULL}.}

\item{gcm_ssp_years}{character. Timeseries years for GCM simulations of future
scenarios specified by \code{ssps}. See \code{\link[=list_gcm_ssp_years]{list_gcm_ssp_years()}} for available years.
Defaults to \code{NULL}.}

\item{gcm_hist_years}{character.  Timeseries years for GCM simulations of the
historical scenario. See \code{\link[=list_gcm_hist_years]{list_gcm_hist_years()}} for available years.
Defaults to \code{NULL}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{...}{other arguments passed to \code{\link[=downscale_core]{downscale_core()}}. Namely: \code{return_normal},
\code{vars}, \code{out_spatial} and \code{plot}}
}
\value{
\code{data.table} of downscaled climate variables for each location.
All outputs are returned in one table.
}
\description{
\code{downscale()} provides downscaled climate variables for user-specified
locations.
\code{downscale()} adapts a simple change-factor (aka "delta') downscaling
approach originally implemented in \href{https://climatena.ca/}{ClimateNA}.
This approach downscales climate data in three stages:
\enumerate{
\item \emph{Change-factor downscaling} of coarse-resolution (50-200km grid) monthly temperature and precipitation data from climate models or observational sources to high-resolution (800m grid);
\item \emph{Elevation adjustment} of temperature variables to provide scales finer than the high-resolution reference grid; and
\item \emph{Calculating derived variables} from the downscaled monthly temperature and precipitation variables.
}
See \code{vignette("methods_downscaling")} for a description of the downscaling methodology.

\code{downscale()} is a user-friendly wrapper for \code{downscale_core()}
}
\details{
\code{\link[=downscale_core]{downscale_core()}} parameters can be applied in \code{climr_downscale()}. For example,
setting \code{ppt_lr = TRUE} in \code{climr_downscale()} will apply elevation adjustment to precipitation values.

Although \code{which_refmap = "auto"} is the default, users are cautioned that
this can produce artefacts associated with downscaling to different reference
climate maps within and outside the western North American boundary of \code{refmap_climr}.
We recommend that queries spanning this boundary use \code{which_refmap = "refmap_climatena"}.
}
\examples{
library(data.table)
library(terra)
set.seed(123)
dbCon <- data_connect()
xyz <- data.frame(
  lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10),
  id = 1:10
)

## get bounding box based on input points
thebb <- get_bb(xyz)
obs <- input_obs(dbCon, thebb, period = "2001_2020")
plot(obs[[1]][[2]])

## provide or create a lon, lat, elev, and optionally id, dataframe - usually read from csv file
in_xyz <- data.frame(
  lon = c(
    -127.70521, -127.62279, -127.56235, -127.7162,
    -127.18585, -127.1254, -126.94957, -126.95507
  ),
  lat = c(55.3557, 55.38847, 55.28537, 55.25721, 54.88135, 54.65636, 54.6913, 54.61025),
  elev = c(291L, 296L, 626L, 377L, 424L, 591L, 723L, 633L),
  id = 1:8,
  Zone = c(rep("CWH", 3), rep("CDF", 5)),
  Subzone = c("vm1", "vm2", "vs1", rep("mm", 3), "dk", "dc")
)

## historic observational time series
vars <- c("PPT", "CMD", "Tave_07")
climate_norms_hist <- climr_downscale(
  xyz = in_xyz, which_normal = "auto",
  return_normal = TRUE,
  obs_periods = "2001_2020",
  vars = vars,
  out_spatial = TRUE, plot = "PPT"
) ## specify desired variables to plot

## as a data.table
climate_norms_hist <- downscale(
  xyz = in_xyz, which_refmap = "auto",
  return_normal = TRUE,
  vars = vars,
  out_spatial = FALSE, plot = "PPT"
) ## specify desired variables to plot

## future projections
climate_norms_fut <- downscale(
  xyz = in_xyz, which_refmap = "auto",
  gcms = c("ACCESS-ESM1-5"),
  ssps = c("ssp370"),
  gcm_periods = c("2021_2040"),
  # gcm_ssp_years = 2020:2060,
  max_run = 3, #' we want 3 individual runs for each model
  vars = vars
)

}
