% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale target rasters to points of interest}
\usage{
downscale(
  xyz,
  normal,
  gcm = NULL,
  vars = sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))
)
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{normal}{Reference normal baseline input from \code{normal_input}.}

\item{gcm}{Global Circulation Models input from \code{gcm_input}. Default to NULL.}

\item{vars}{A character vector of climate variables to compute. Supported variables
can be obtained with \code{list_variables()}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}
}
\value{
A downscaled dataset. If \code{gcm} is NULL, this is just the downscaled \code{normal}
at point locations. If \code{gcm} is provided, this returns a downscaled dataset for each
point location, general circulation model, shared socioeconomic pathway, run and period.
}
\description{
Downscale target rasters to points of interest
}
\examples{
\dontrun{
xyz <- data.frame(lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10))
normal <- normal_input()
gcm_input <- gcm_input(list_gcm()[3], list_ssp()[1], list_period()[2])
downscale(xyz, normal, gcm)
}
}
