% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale target rasters to points of interest}
\usage{
downscale(
  xyz,
  historical,
  future = NULL,
  variables,
  grouping = c("m", "s", "a")
)
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{historical}{Reference historical climate variables rasters.}

\item{future}{Future climate variables rasters. Default to NULL.}

\item{variables}{A character vector of variables to compute.}

\item{grouping}{A character vector of variables grouping. Can be \code{m} (monthly), \code{s} (seasonal) or \code{a} (annual).}
}
\description{
Downscale target rasters to points of interest
}
