% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale target rasters to points of interest}
\usage{
downscale(xyz, normal, gcm = NULL, extra = character())
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{normal}{Reference normal baseline input from \code{normal_input}.}

\item{gcm}{Global Circulation Models input from \code{gcm_input}. Default to NULL.}

\item{extra}{A character vector of extra variables to compute. Supported variables
can be obtained with \code{list_variables(FALSE)}. Definitions can be found in this package
\code{variables} dataset. Default to \code{character()}.}
}
\description{
Downscale target rasters to points of interest
}
\examples{
\dontrun{
xyz <- data.frame(lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10))
normal <- normal_input()
gcm_input <- gcm_input(list_gcm()[3], list_ssp()[1], list_period()[2])
downscale(xyz, normal, gcm)
}
}
