% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Self-contained change-factor downscaling of observed and simulated climate data}
\usage{
downscale(
  xyz,
  which_refmap = "auto",
  obs_periods = NULL,
  obs_years = NULL,
  obs_ts_dataset = NULL,
  gcms = NULL,
  ssps = NULL,
  gcm_periods = NULL,
  gcm_ssp_years = NULL,
  gcm_hist_years = NULL,
  max_run = 0L,
  cache = TRUE,
  ...
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{which_refmap}{character. Which map of 1961-1990 climatological normals to use as the
high-resolution reference climate map for downscaling. Default is "auto",
which selects, for each query point, the best available climatological
normals map in declining order of \code{"refmap_prism"}, \code{"refmap_climr"}, and \code{"refmap_climatena"}.
Other options are one of \code{\link[=list_refmaps]{list_refmaps()}}, which will provide a consistent
reference map for all points.}

\item{obs_periods}{character. Which historical period for observational climate
data, averaged over this period. Options are \code{\link[=list_obs_periods]{list_obs_periods()}}. Default \code{NULL}}

\item{obs_years}{integer. Vector of years to obtain individual years or time
series of observational climate data. Default \code{NULL}. See \code{\link[=list_obs_years]{list_obs_years()}}
for available years.}

\item{obs_ts_dataset}{character. The dataset to use for observational time series data. Options
are \code{"climatena"} for the ClimateNA gridded time series or \code{"cru.gpcc"} for the combined Climatic
Research Unit TS dataset (for temperature) and Global Precipitation Climatology Centre dataset
(for precipitation). Defaults to \code{NULL}.}

\item{gcms}{character. Vector of global climate model names. Options
are \code{\link[=list_gcms]{list_gcms()}}. Defaults to \code{NULL}.}

\item{ssps}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssps]{list_ssps()}}. Defaults to all scenarios available.}

\item{gcm_periods}{character. 20-year reference periods for GCM simulations.
Options are \code{\link[=list_gcm_periods]{list_gcm_periods()}}. Defaults to \code{NULL}.}

\item{gcm_ssp_years}{character. Timeseries years for GCM simulations of future
scenarios specified by \code{ssps}. See \code{\link[=list_gcm_ssp_years]{list_gcm_ssp_years()}} for available years.
Defaults to \code{NULL}.}

\item{gcm_hist_years}{character.  Timeseries years for GCM simulations of the
historical scenario. See \code{\link[=list_gcm_hist_years]{list_gcm_hist_years()}} for available years.
Defaults to \code{NULL}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{cache}{logical. Cache data locally? Default \code{TRUE}}

\item{...}{other arguments passed to \code{\link[=downscale_core]{downscale_core()}}. Namely: \code{return_refperiod},
\code{vars}, \code{out_spatial} and \code{plot}}
}
\value{
\code{data.table} of downscaled climate variables for each location.
All outputs are returned in one table.
}
\description{
\code{downscale()} provides downscaled climate variables for user-specified
locations.
It adapts a simple change-factor (aka "delta') downscaling
approach originally implemented in \href{https://climatena.ca/}{ClimateNA}.
This approach downscales climate data in three stages:
\enumerate{
\item \emph{Change-factor downscaling} of coarse-resolution (50-200km grid) monthly temperature and precipitation data from climate models or observational sources to high-resolution (800m grid);
\item \emph{Elevation adjustment} of temperature variables to provide scales finer than the high-resolution reference grid; and
\item \emph{Calculating derived variables} from the downscaled monthly temperature and precipitation variables.
}
See \code{vignette("methods_downscaling")} for a description of the downscaling methodology.

\code{downscale()} is a user-friendly wrapper for \code{downscale_core()}
}
\details{
\code{\link[=downscale_core]{downscale_core()}} parameters can be applied in \code{downscale()}. For example,
setting \code{ppt_lr = TRUE} in \code{downscale()} will apply elevation adjustment to precipitation values.

Although \code{which_refmap = "auto"} is the default, users are cautioned that
this can produce artefacts associated with downscaling to different reference
climate maps within and outside the western North American boundary of \code{refmap_climr}.
We recommend that queries spanning this boundary use \code{which_refmap = "refmap_climatena"}.
}
\examples{

## provide or create a dataframe of lon, lat, elev, and id - usually read from csv file
in_xyz <- data.frame(
  lon = c(-127.7052, -127.6227, -127.5623, -127.7162, -127.1858, -127.125, -126.9495, -126.9550),
  lat = c(55.3557, 55.38847, 55.28537, 55.25721, 54.88135, 54.65636, 54.6913, 54.61025),
  elev = c(291, 296, 626, 377, 424, 591, 723, 633),
  id = 1:8
)

## historic observational time series
vars <- c("PPT", "CMD", "Tave_07")
climate_norms_hist <- downscale(
  xyz = in_xyz,
  which_refmap = "auto",
  return_refperiod = TRUE,
  obs_periods = "2001_2020",
  vars = vars,
  out_spatial = TRUE, plot = "PPT"
) ## specify desired variables to plot

## as a data.table
climate_norms_hist <- downscale(
  xyz = in_xyz, which_refmap = "auto",
  return_refperiod = TRUE,
  vars = vars,
  out_spatial = FALSE, plot = "PPT"
) ## specify desired variables to plot

## future projections for annual variables from three models
climate_norms_fut <- downscale(
  xyz = in_xyz, which_refmap = "auto",
  gcms = list_gcms()[c(1, 5, 6)],
  ssps = list_ssps()[2],
  gcm_periods = list_gcm_periods()[1:2],
  # gcm_ssp_years = 2020:2060,
  max_run = 3, #' we want 3 individual runs for the model
  vars = list_vars("Annual")
)

}
