% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale target rasters to points of interest}
\usage{
downscale(
  xyz,
  normal,
  gcm = NULL,
  historic = NULL,
  gcm_ts = NULL,
  gcm_hist = NULL,
  historic_ts = NULL,
  return_normal = FALSE,
  vars = sort(sprintf(c("PPT\%02d", "Tmax\%02d", "Tmin\%02d"), sort(rep(1:12, 3)))),
  ppt_lr = FALSE,
  nthread = 1L
)
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{normal}{Reference normal baseline input from \code{normal_input}.}

\item{gcm}{Global Circulation Models input from \code{gcm_input}. Default to NULL.}

\item{historic}{Historic time period input from \code{historic_input}. Default to NULL}

\item{gcm_ts}{TODO}

\item{gcm_hist}{TODO}

\item{historic_ts}{TODO}

\item{return_normal}{TODO}

\item{vars}{A character vector of climate variables to compute. Supported variables
can be obtained with \code{list_variables()}. Definitions can be found in this package
\code{variables} dataset. Default to monthly PPT, Tmax, Tmin.}

\item{ppt_lr}{A boolean. Apply lapse rate adjustment to precipitations. Default to FALSE.}

\item{nthread}{An integer. Number of parallel threads to use to do computations. Default to 1L.}
}
\value{
A \code{data.table} with downscaled climate variables. If \code{gcm} is NULL,
this is just the downscaled \code{normal} at point locations. If \code{gcm} is provided,
this returns a downscaled dataset for each point location, general circulation
model (GCM), shared socioeconomic pathway (SSP), run and period.
}
\description{
Downscale target rasters to points of interest
}
\details{
Additional details... TODO.
}
\examples{
\dontrun{
dbCon <- data_connect()
xyz <- data.frame(lon = runif(10, -140, -106), lat = runif(10, 37, 61), elev = runif(10))

## get bounding box based on input points
thebb <- get_bb(xyz)
normal <- normal_input(dbCon = dbCon, bbox = thebb, cache = TRUE)

## pick one GCM, one SSP and one period from the list of available options 
gcm <- gcm_input(dbCon, thebb, gcm = list_gcm()[3], list_ssp()[1], list_gcm_period()[2])

## notice coarseness of the data
terra::plot(gcm[[1]])

downscale(xyz, normal, gcm)
historic <- historic_input(dbCon, thebb)
terra::plot(historic[[1]])

downscale(xyz, normal, gcm = NULL, historic = historic, ppt_lr = FALSE)
}
}
