% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_db}
\alias{extract_db}
\title{Bilinear interpolation point extraction from raster bands}
\usage{
extract_db(
  rastertbl,
  layers = data.table::data.table(var_nm = character(), laynum = integer()),
  VAR = NULL,
  hull = NULL
)
}
\arguments{
\item{rastertbl}{The name of the raster table to extract from.}

\item{layers}{A data.table with column \code{var_nm} (names of the bands) and}

\item{VAR}{In case of rasters split across different tables. One VAR in each.
The first VAR will be used for replacement in labels names. Assuming all
tables have the same structure, only different VAR in them.}

\item{hull}{An optional hull well-known text geometry to prefilter
raster table before merging points projected in srid 4326.
\code{laynum} (index number of the bands to extract).}
}
\value{
A data.frame one row per point id in xyz with an ID
column + one columns for each \code{bands} in \code{rastertbl} name \code{nm}.
}
\description{
Bilinear interpolation point extraction from raster bands
}
\keyword{internal}
