% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_climr.R
\name{fit}
\alias{fit}
\title{Fit basic statistical models to climate data}
\usage{
fit(
  obj,
  data_type = c("yearly", "quarterly", "monthly"),
  fit_type = c("lm", "loess", "smooth.spline")
)
}
\arguments{
\item{obj}{An object of class \code{"climr"} generated by \code{\link{load_climr}}.}

\item{data_type}{The type of data to be modelled, either \code{"yearly"}
(the default), \code{"quarterly"}, or \code{"monthly"}.}

\item{fit_type}{The type of model required, either linear regression (\code{"lm"}) by default,
\code{"loess"}, or smoothing splines (\code{"smooth.spline"}).}
}
\value{
An object of class \code{"climr_fit"} which includes the model details
as well as the data set and \code{fit_type} used.
}
\description{
Fits linear regression, LOESS, or smoothing splines to hemispheric data from NASA.
}
\note{
A dedicated \code{\link{plot.climr_fit}} method is provided
for objects of class \code{"climr_fit"}
.
}
\examples{
dat <- load_climr(type = "SH")
mod1 <- fit(dat)
mod2 <- fit(dat, data_type="monthly", fit_type="smooth.spline")
mod3 <- fit(dat, data_type="quarterly", fit_type="loess")
}
\seealso{
\code{\link{load_climr}}, \code{\link{plot.climr_fit}}, \code{\link[stats]{lm}},
\code{\link[stats]{loess}}, \code{\link[stats]{smooth.spline}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@mu.ie}>
}
