% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{gcm_input}
\alias{gcm_input}
\alias{gcm_hist_input}
\alias{gcm_ts_input}
\title{Create GCM inputs for \code{downscale} using data on Postgis database.}
\usage{
gcm_input(
  dbCon,
  bbox = NULL,
  gcm = list_gcm(),
  ssp = list_ssp(),
  period = list_gcm_period(),
  max_run = 0L,
  cache = TRUE
)

gcm_hist_input(
  dbCon,
  bbox = NULL,
  gcm = list_gcm(),
  years = 1901:1950,
  max_run = 0L,
  cache = TRUE
)

gcm_ts_input(
  dbCon,
  bbox = NULL,
  gcm = list_gcm(),
  ssp = list_ssp(),
  years = 2020:2030,
  max_run = 0L,
  cache = TRUE
)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{Numeric vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{gcm}{A character vector. Label of the global circulation models to use.
Can be obtained from \code{list_gcm()}. Default to \code{list_gcm()}.}

\item{ssp}{A character vector. Label of the shared socioeconomic pathway(s) to use.
See available options with \code{list_ssp()}. Defaults to \code{list_ssp()}.}

\item{period}{A character vector. Label of the period to use.
Can be obtained from \code{list_gcm_period()}. Default to \code{list_gcm_period()}.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{cache}{Logical specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}

\item{years}{Numeric or character vector in \code{2020:2100}. Defaults to \code{2020:2030}.}
}
\value{
A \code{list} of SpatRasters, each with possibly multiple layers, that can
be used with \code{downscale}.
}
\description{
\code{gcm_input} creates GCM climate periods inputs, given chosen GCMs, SSPs,
periods and runs.

\code{gcm_hist_input} creates GCM \strong{historic} time series inputs, given chosen GCMs, SSPs,
years and runs.

\code{gcm_ts_input} creates GCM time series inputs, given chosen GCMs, SSPs,
years and runs.
}
