% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bb}
\alias{get_bb}
\title{Find bounding box of data}
\usage{
get_bb(in_xyz)
}
\arguments{
\item{in_xyz}{\code{data.table} (or \code{data.frame}) of points to downscale
with columns "lon", "lat", "elev" and "id"}
}
\value{
numeric vector. Bounding box coordinates with order ymax,ymin,xmax,xmin (e.g. \code{c(51, 50, -121, -122)}).
}
\description{
Find bounding box of data
}
