% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fit.climr.R
\name{gp_fit}
\alias{gp_fit}
\title{Fit Gaussian Process and creates predictions on climate data using optim}
\usage{
gp_fit(obj, method = c("Nelder-Mead", "BFGS", "SANN", "Brent"))
}
\arguments{
\item{obj}{An object of class \code{climr} from \code{\link{load_clim}}}

\item{method}{The type of optimisation method required, either Nelder-Mead (\code{Nelder-Mead}), BFGS(\code{BFGS}),SANN(\code{SANN}) or Brent (\code{Brent})}
}
\value{
Returns a list of class \code{climr_gp_fit} which includes the predicted values as well as the data set and method type used
}
\description{
Fit Gaussian Process and creates predictions on climate data using optim
}
\examples{
data1 = load_clim('NH')
data2 = gp_fit(data1, 'BFGS')
}
\seealso{
\code{\link{load_clim}}, \code{\link{plot.climr_gp_fit}},\code{\link{plot.climr_fit}}, \code{\link{fit}}
}
