% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fit.climr.R
\name{gp_fit}
\alias{gp_fit}
\title{Fit basic statistical models to climate data}
\usage{
gp_fit(obj, optim_method = c("Nelder-Mead", "BFGS", "SANN", "Brent"))
}
\arguments{
\item{obj}{An object of class \code{climr} from \code{\link{load_clim}}}

\item{optim_method}{The optimisation algorithm to be used}
}
\value{
Returns a list of class \code{gp_fit} which includes the model details as well as the data set and fit type used
}
\description{
Fit basic statistical models to climate data
}
\examples{
ans1 = load_clim('SH')
ans2 = gp_fit(ans1, optim_method = "BFGS")
}
\seealso{
\code{\link{load_clim}}, \code{\link{plot.climr_gp_fit}}
}
