% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{input_obs}
\alias{input_obs}
\alias{input_obs_ts}
\title{Retrieve observational anomalies.}
\usage{
input_obs(period = list_obs_periods(), cache = TRUE)

input_obs_ts(
  dataset = c("cru.gpcc", "climatena"),
  years = 2010:2022,
  cache = TRUE
)
}
\arguments{
\item{period}{character. Vector of labels of the periods to use.
Can be obtained from \code{\link[=list_obs_periods]{list_obs_periods()}}. Default to "2001_2020".}

\item{cache}{logical. Specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}

\item{dataset}{Character. Which observational dataset to use? Options are \code{"climatena"} for the
ClimateNA gridded time series or \code{"cru.gpcc"} for the combined Climatic Research Unit TS dataset
(for temperature) and Global Precipitation Climatology Centre dataset (for precipitation).}

\item{years}{numeric. Years to retrieve obs anomalies for. Defaults to \code{2010:2022}.
See \code{\link[=list_obs_years]{list_obs_years()}} for available years.}
}
\value{
An input list, with two nodes, \code{tbl} and \code{layers}, suitable for \code{extract_rast_tbl}.

An input list, with two nodes, \code{tbl} and \code{layers}, suitable for \code{extract_rast_tbl}.
}
\description{
\code{input_obs} produces anomalies of the average observed climate for a given \strong{period},
relative to the 1961-1990 reference period. The anomalies are calculated from the \code{"cru.gpcc"} dataset
which is the Climatic Research Unit TS dataset (for temperature) and Global Precipitation Climatology Centre dataset
(for precipitation).

\code{input_obs_ts} produces anomalies of observed climate for a given \strong{time series} of individual years.
}
\details{
Generally, this function should only be used in combination with \code{\link[=downscale_core]{downscale_core()}} as the values
returned in the rasters are anomalies compared to the 1961-1990 reference period,
and are usually not meaningful without the whole downscale process.

The returned raster contains anomalies for each year specified in \code{years}. In general this
function should only be used in conjunction with \code{\link[=downscale_core]{downscale_core()}}.
}
\seealso{
\code{\link[=downscale_core]{downscale_core()}}, \code{\link[=list_obs_periods]{list_obs_periods()}}
}
