% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_input}
\alias{historic_input}
\title{Create historic input for \code{downscale}.}
\usage{
historic_input(dbCon, bbox = NULL, period = list_historic(), cache = TRUE)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{Numeric vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{period}{A character vector. Label of the period to use.
Can be obtained from \code{list_gcm_period()}. Default to \code{list_gcm_period()}.}

\item{cache}{Logical specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
A \code{list} of SpatRasters, each with possibly multiple layers, that can
be used with \code{downscale}.
}
\description{
Create historic input for \code{downscale}.
}
