% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_input_ts}
\alias{historic_input_ts}
\title{Create historic timeseries input for \code{downscale}.}
\usage{
historic_input_ts(dbCon, bbox = NULL, years = 2010:2022, cache = TRUE)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{Numeric vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{years}{Years to retrieve timeseries for, in \code{1902:2022}. Default \code{2010:2022}}

\item{cache}{Logical specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
A \code{list} of SpatRasters, each with possibly multiple layers, that can
be used with \code{downscale}.
}
\description{
Create historic timeseries input for \code{downscale}.
}
