% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_in_bbox}
\alias{is_in_bbox}
\title{Check that new BB is in old BB}
\usage{
is_in_bbox(newbb, oldbb)
}
\arguments{
\item{newbb}{TODO}

\item{oldbb}{TODO}
}
\value{
logical
}
\description{
Check that new BB is in old BB
}
