% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate}
\alias{lapse_rate}
\title{Lapse rate computation}
\usage{
lapse_rate(
  dem,
  targets,
  replace_NaN = TRUE,
  as_raster = FALSE,
  use_parallel = FALSE
)
}
\arguments{
\item{dem}{A digital elevation model rasterLayer.}

\item{targets}{A list of target raster Layers to compute lapse rates for.}

\item{replace_NaN}{A boolean. Should NaN lapse rates be replaced by zeros. Default to TRUE.}

\item{as_raster}{A boolean. Should the results be returned as a list of raster. Default to FALSE.}

\item{use_parallel}{A boolean. Should parallel::mclapply be used. Default to FALSE.
with the same extend as the digital elevation model.}
}
\value{
Lapse rate values.
}
\description{
Lapse rate computation
}
\details{
Formulas
Simple linear regression without the intercept term
β = Σxy / Σx²
mss = Σ(xβ)², sum of squared fitted values
rss = Σε², sum of squared (y minus fitted), sum of absolute errors
R² = mss / (mss + rss)
Lapse rate = βR²
}
