% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate}
\alias{lapse_rate}
\title{Lapse rate computation}
\usage{
lapse_rate(target, NA_replace = TRUE, use_parallel = TRUE, rasterize = TRUE)
}
\arguments{
\item{target}{Target rasters to compute lapse rates for. Build with this package functions.}

\item{NA_replace}{A boolean. Should NA lapse rate results be replaced by zeros. Default to TRUE.}

\item{use_parallel}{A boolean. Should parallel::mclapply be used. Default to TRUE.}

\item{rasterize}{Return an object of the same class category as target with the same extend.}
}
\value{
Lapse rate values.
}
\description{
Lapse rate computation
}
\details{
Formulas
Simple linear regression without the intercept term
β = Σxy / Σx²
mss = Σ(xβ)², sum of squared fitted values
rss = Σε², sum of squared (y minus fitted), sum of absolute errors
R² = mss / (mss + rss)
Lapse rate = βR²
}
