% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate}
\alias{lapse_rate}
\title{Calculation of gridded lapse rates}
\usage{
lapse_rate(normal, dem, NA_replace = TRUE, nthread = 1L, rasterize = TRUE)
}
\arguments{
\item{normal}{a \code{SpatRaster} stack. climate rasters to compute lapse rates for. Build with this package functions.}

\item{dem}{\code{SpatRaster}. A digital elevation model raster. Only needed if \code{normal}
is a \code{SpatRaster}. Defaults to \code{NULL}.}

\item{NA_replace}{logical. Should NA lapse rate results be replaced by zeros. Default to TRUE.}

\item{nthread}{integer. Number of parallel threads to use to compute lapse rates.}

\item{rasterize}{logical. Return an object of the same class category as normal with the same extent?
Default to \code{TRUE}.}
}
\value{
\code{SpatRaster} of lapse rate values.
}
\description{
This function computes lapse rates (a linear relationship of a climate variable to elevation) across user-provided rasters.
Lapse rates are calculated using simple linear regression (without the intercept) using the focal cell and its eight neighbors as observations.
These gridded lapse rates are used in \code{\link[=downscale]{downscale()}} and \code{\link[=climr_downscale]{climr_downscale()}} for elevation adjustment of climate values during downscaling to user-specified locations.
The method is illustrated in the vignette 'vignette("lapse_rates")'
}
\details{
Formulas \cr
beta_coef = sum(xy) / sum(x²) \cr
mss = sum(x * beta_coef)², sum of squared fitted values \cr
rss = sum(ε²), sum of squared (y minus fitted), sum of absolute errors \cr
R² = mss / (mss + rss) \cr
Lapse rate = beta_coef * R²
}
