% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapserate.R
\name{lapse_rate_redux}
\alias{lapse_rate_redux}
\title{TODO: add documentation here
For the lapse rate, x is the elevation, and y is the normal}
\usage{
lapse_rate_redux(y_i, x_i, n_r, n_c, n_sc, sum_xx, NA_replace)
}
\arguments{
\item{y_i}{TODO}

\item{x_i}{TODO}

\item{n_r}{TODO}

\item{n_c}{TODO}

\item{n_sc}{TODO}

\item{sum_xx}{TODO}

\item{NA_replace}{TODO}
}
\value{
\code{data.table}
}
\description{
TODO: add documentation here
For the lapse rate, x is the elevation, and y is the normal
}
