% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_ramp.R
\name{legend_ramp}
\alias{legend_ramp}
\title{Legend for color ramp maps}
\usage{
legend_ramp(
  r,
  title,
  ColScheme,
  breaks,
  pos = c(0.2, 0.23, 0.1, 0.5),
  log = NULL,
  horizontal = FALSE,
  title.height = 1
)
}
\arguments{
\item{r}{A spatial raster object, typically from the \code{terra} package.}

\item{title}{Character. The title for the legend.}

\item{ColScheme}{Character vector. Specifies the colors used for the legend. Typically corresponds
to the color scheme applied to the raster data visualization.}

\item{breaks}{Numeric vector. Defines the breakpoints for the legend, determining how the range
of the raster is divided into intervals corresponding to the colors in \code{ColScheme}.}

\item{pos}{Numeric vector of length 4. Defines the position of the legend as
proportions of the raster extent. The format is \code{c(xmin, xmax, ymin, ymax)}.}

\item{log}{Numeric or NULL. The base of the logarithm used to compute legend labels.
If \code{NULL}, values are displayed as-is.}

\item{horizontal}{Logical. Should the legend be drawn horizontally? Default is \code{FALSE}.}

\item{title.height}{Numeric. Scaling factor for the height of the legend title relative to the legend size.}
}
\value{
NULL. A plot of the legend is drawn in the current plotting device. The function does not return any value.
}
\description{
This function plots a map legend that handles logarithmic scaling
}
\details{
The function uses the extent of the raster object to determine the positioning and size of
the legend. The legend can be oriented either horizontally or vertically, and logarithmic
scaling is applied to the labels if the \code{log} parameter is specified.
}
\examples{
dem <- get(data("dem_vancouver")) |> unwrap() ## get the sample digital elevation model (dem) provided with `climr`
grid <- as.data.frame(dem, cells = TRUE, xy = TRUE) ## convert the DEM to a data.frame
colnames(grid) <- c("id", "lon", "lat", "elev") # rename column names to what climr expects
ds_out <- downscale(grid, which_refmap = "refmap_prism") ## A simple climr query. This will return the observed 1961-1990 normals for the raster grid points. 
clim <- rast(dem) # use the DEM as a template raster
clim[ds_out[, id]] <- ds_out[, get(var)] # populate the raster cells with the 2001-2020 annual precipitation (MAP) values, using the `id` field as the link. 
clim <- log2(clim) # log-transform precipitation for more meaningful scaling
inc=diff(range(values(clim)))/500 #increment for the ramp
breaks=seq(min(values(clim))-inc, max(values(clim))+inc, inc) # color breaks
ColScheme <- rev(hcl.colors(length(breaks)-1, "GnBu")) # color scheme
plot(clim, col=ColScheme, breaks=breaks, legend=F, main="", mar=NA)
legend_ramp(clim, title = paste(var, "(mm)"), ColScheme = ColScheme, breaks = breaks, pos=c(0.05, 0.45, 0.1, 0.125), log = 2, horizontal = TRUE)

}
