% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{list_run}
\alias{list_run}
\title{List available runs}
\usage{
list_run(gcm)
}
\arguments{
\item{gcm}{An optional character vector. Limit list to provided global circulation models.}
}
\description{
List available runs
}
