% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{list_run}
\alias{list_run}
\title{List available runs for given GCM}
\usage{
list_run(dbCon, gcm)
}
\arguments{
\item{dbCon}{database connection}

\item{gcm}{Character vector to specify requested GCMs}
}
\description{
List available runs for given GCM
}
