% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{list_variables}
\alias{list_variables}
\title{List climate variables}
\usage{
list_variables(only_extra = TRUE)
}
\arguments{
\item{only_extra}{A boolean. Should Tmin, Tmax and PPT be excluded? Default to TRUE.}
}
\description{
List climate variables
}
