% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_clim_vars.R
\name{list_variables}
\alias{list_variables}
\title{List climate variables}
\usage{
list_variables(only_extra = FALSE)
}
\arguments{
\item{only_extra}{A boolean. Should Tmin, Tmax and PPT be excluded? Defaults to FALSE.}
}
\value{
a character vector
}
\description{
List climate variables
}
