% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{make_raster}
\alias{make_raster}
\title{Make raster from a boundary}
\usage{
make_raster(boundary)
}
\arguments{
\item{boundary}{Numeric vector of length 4 (ymax,ymin,xmax,xmin). Defaults to \code{NULL}.}
}
\value{
a SpatRaster
}
\description{
Used internally to access the PostGRS database and
create a SpatRaster using a given spatial boundary
}
