% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetRaster.R
\name{make_raster}
\alias{make_raster}
\title{Make raster from a boundary}
\usage{
make_raster(boundary, conn, rastque, nameque, projID, bands)
}
\arguments{
\item{boundary}{numeric. Vector of length 4 (ymax,ymin,xmax,xmin). Defaults to \code{NULL}.}

\item{conn}{a DBI or RPostgres connection object.}

\item{rastque}{character. The "rast" query name obtained with e.g. \code{dbQuoteIdentifier(conn, "rast")}}

\item{nameque}{character. The "schema.name"}

\item{projID}{character. projID in data.base}

\item{bands}{numeric. Which raster bands to return. Default 37:73.}
}
\value{
a \code{SpatRaster}
}
\description{
Used internally to access the PostGRS database and
create a \code{SpatRaster} using a given spatial boundary
}
