% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_input}
\alias{normal_input}
\title{Create normal baseline inputs for \code{downscale} from PostGIS database.}
\usage{
normal_input(dbCon, bbox = NULL, normal = "normal_na", cache = TRUE)
}
\arguments{
\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{bbox}{numeric. Vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{normal}{A \code{character} or a \code{SpatRaster}. For character, label of the normal baseline to use.
Can be obtained from \code{list_normal()}. For \code{SpatRaster}, 36 layers normal climate variables with
names PPT01:PPT12, Tmax01:Tmax12 and Tmin01:Tmin12. Defaults to \code{list_normal()[1]}.}

\item{cache}{logical. Specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
A \code{SpatRaster} containing normals, lapse rates
and digital elevation model layers, that can be used with \code{downscale}.
}
\description{
Create normal baseline inputs for \code{downscale} from PostGIS database.
}
