% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{packRasters}
\alias{packRasters}
\title{Pack rasters for parallel computing}
\usage{
packRasters(ras)
}
\arguments{
\item{ras}{a \code{SpatRaster} or list of \code{SpatRasters}, or \code{NULL}}
}
\value{
\code{NULL}, a packed \code{SpatRaster} or list of packed \code{SpatRasters}
}
\description{
Pack rasters for parallel computing
}
