% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climr.R
\name{plot.climr_fit}
\alias{plot.climr_fit}
\title{Plot climr output}
\usage{
\method{plot}{climr_fit}(x, time_grid = pretty(x$data$x, n = 100), ...)
}
\arguments{
\item{x}{An object of class \code{"climr_fit"} outputted from the \code{link{fit}} function.}

\item{time_grid}{An optional time grid over which to produce fitted values of the model.}

\item{...}{Catches unused arguments to \code{plot} (not currently implemented).}
}
\value{
Nothing, just a nice ggplot.
}
\description{
Loads in global or hemispheric data from NASA at different intervals.
}
\examples{
dat <- load_climr(type = "SH")
mod1 <- fit(dat)
mod2 <- fit(dat, data_type="monthly", fit_type="smooth.spline")
mod3 <- fit(dat, data_type="quarterly", fit_type="loess")
plot(mod1)
plot(mod2)
plot(mod3)
}
\seealso{
\code{\link{fit}}, \code{\link{load_climr}}
}
\author{
Keefe Murphy - <\email{keefe.murphy@mu.ie}>
}
