% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.climr_gp_fit.R
\name{plot.climr_gp_fit}
\alias{plot.climr_gp_fit}
\title{Plot climr_gp_fit output}
\usage{
\method{plot}{climr_gp_fit}(x, time_grid = pretty(x$data$year, n = 100), ...)
}
\arguments{
\item{x}{Output from the \code{\link{gp_fit}} function}

\item{time_grid}{An optional time grid over which to produce fitted values of the model}

\item{...}{Other arguments to plot (not currently implemented)}
}
\value{
A nice plot to illustrate the output of gp_fit as a scatterplot, and the smoothed Gaussian process regression line
}
\description{
Plot climr_gp_fit output
}
\examples{
data1 = load_clim('NH')
data2 = gp_fit(data1, 'BFGS')
plot(data2)
}
\seealso{
\code{\link{load_clim}}, \code{\link{gp_fit}}, \code{\link{plot.climr_fit}}, \code{\link{fit}}
}
