% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.climr_gp_fit.R
\name{plot.climr_gp_fit}
\alias{plot.climr_gp_fit}
\title{Plot climr output}
\usage{
\method{plot}{climr_gp_fit}(x, ...)
}
\arguments{
\item{x}{Output from the \code{\link{gp_fit}} function}

\item{...}{Other arguments to plot (not currently implemented)}
}
\value{
Nothing: just a nice plot
}
\description{
Plot climr output
}
\examples{
ans1 = load_clim('SH')
ans2 = gp_fit(ans1, optim_method = "BFGS")
plot(ans2)
}
\seealso{
\code{\link{load_clim}}, \code{\link{gp_fit}}
}
