% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries.R
\name{plot.ensemble}
\alias{plot.ensemble}
\title{Function for plotting time series for gcms or compiled ensemble}
\usage{
plot.ensemble(
  x,
  var,
  scenarios.selected,
  scenarios,
  showrange = TRUE,
  simplify = TRUE,
  gcm,
  pal,
  pal.scenario,
  refline = FALSE,
  showmean = TRUE,
  endlabel = "change",
  element,
  compile = TRUE,
  var2 = NULL,
  element1,
  element2,
  yeartime.names,
  yeartimes,
  yeartime
)
}
\arguments{
\item{x}{climate data}

\item{var}{TODO}

\item{scenarios.selected}{TODO}

\item{scenarios}{TODO}

\item{showrange}{TODO}

\item{simplify}{TODO}

\item{gcm}{TODO}

\item{pal}{TODO}

\item{pal.scenario}{TODO}

\item{refline}{TODO}

\item{showmean}{TODO}

\item{endlabel}{TODO}

\item{element}{TODO}

\item{compile}{TODO}

\item{var2}{TODO}

\item{element1}{TODO}

\item{element2}{TODO}

\item{yeartime.names}{TODO}

\item{yeartimes}{TODO}

\item{yeartime}{TODO}
}
\description{
Function for plotting time series for gcms or compiled ensemble
}
