% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries.R
\name{plot.ensemble}
\alias{plot.ensemble}
\title{function for plotting time series for gcms or compiled ensemble}
\usage{
plot.ensemble(x)
}
\description{
function for plotting time series for gcms or compiled ensemble
}
