% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_diagram.R
\name{plot_climate_diagram}
\alias{plot_climate_diagram}
\title{Plot Climate Diagram}
\usage{
plot_climate_diagram(temp, precip, elev = NULL)
}
\arguments{
\item{temp}{vector of temperatures (in Celsius) for each month}

\item{precip}{vector of precipitation (in mm) for each month}

\item{elev}{elevation (in m)s}
}
\description{
Plot Climate Diagram
}
\examples{

temperature <- c(14, 14, 17, 20, 24, 28, 31, 31, 27, 23, 18, 15)
precipitations <- c(93.1, 62.1, 58.8, 46.6, 45.9, 23.1, 11.2, 18.9, 73.8, 68, 155.3, 76.3)
plot_climate_diagram(temp = temperature, precip = precipitations,  elev = 25)

plot_climate_diagram(temp = c(-3, 14, 17, 20, 24, 28, 31, 31, 27, 23, 18, -5), 
   precip = c(93.1, 62.1, 58.8, 46.6, 45.9, 23.1, 11.2, 18.9, 73.8, 68, 155.3, 76.3), 
   elev = 25 )

}
