% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeSeries_input.R
\name{plot_timeSeries_input}
\alias{plot_timeSeries_input}
\title{Input data for the time series climate change plot}
\usage{
plot_timeSeries_input(
  xyz,
  gcm_models = list_gcm(),
  ssp = list_ssp(),
  max_run = 10,
  historic_ts_dataset = "cru.gpcc",
  historic_ts = 1901:2022,
  gcm_hist_years = 1850:2014,
  gcm_ts_years = 2015:2100,
  vars = list_variables()
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{gcm_models}{character. Vector of global climate model names. Options
are \code{\link[=list_gcm]{list_gcm()}}. Used for gcm periods, gcm timeseries, and historic timeseries.
Defaults to \code{NULL}.}

\item{ssp}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssp]{list_ssp()}}. Defaults to all scenarios available.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{historic_ts_dataset}{character. The dataset to use for observed time series data. Options
are \code{"climate_na"} for the ClimateNA gridded time series or \code{"cru_gpcc"} for the combined Climatic
Research Unit TS dataset (for Temperature) and Global Precipitation Climatology Centre dataset
(for precipitation). Defaults to \code{NULL}.}

\item{historic_ts}{integer. Vector of years to obtain individual years or time
series of observed climate data. Default \code{NULL}. See \code{\link[=list_historic_ts]{list_historic_ts()}}
for available years.}

\item{gcm_hist_years}{character.  Timeseries years for GCM simulations of the
historical scenario. See \code{\link[=list_gcm_hist_ts]{list_gcm_hist_ts()}} for available years.
Defaults to \code{NULL}.}

\item{gcm_ts_years}{character. Timeseries years for GCM simulations of future
scenarios specified by \code{ssp}. See \code{\link[=list_gcm_ts]{list_gcm_ts()}} for available years.
Defaults to \code{NULL}.}
}
\value{
\code{data.table} of average downscaled climate variables for all locations.
}
\description{
Input data for the \code{\link[=plot_timeSeries]{plot_timeSeries()}} function. Since these inputs are time-consuming to generate,
the purpose of conducting the generation of the input table in a separate function is to allow users
to make multiple calls to \code{\link[=plot_timeSeries]{plot_timeSeries()}} (e.g., for comparing different climate variables)
without needing to generate the inputs each time.
}
\details{
This function generates standardized inputs for one or multiple locations at any spatial scale.
If multiple locations are specified, the output is the average of the climate variables for all locations.
}
\examples{
# data frame of arbitrary points
my_points <- data.frame(lon = c(-127.7300,-127.7500), lat = c(55.34114, 55.25), elev = c(711, 500), id = 1:2)

# generate the input data
my_data <- plot_timeSeries_input(my_points)

# use the input to create a plot
plot_timeSeries(my_data, variable1 = "Tmin_sm")

}
