% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_cache.R
\name{pre_cache}
\alias{pre_cache}
\title{Pre cache data for presets or custom areas}
\usage{
pre_cache(region = c("BC", "WNA", "NA"), bbox = NULL)
}
\arguments{
\item{region}{character. Preset region; options are one of "BC", "WNA", and "NA"}

\item{bbox}{optional custom bounding box (or SpatExtent) for region of interest. If bbox is specified, region will be ignored.}
}
\description{
TODO
}
