% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm.R
\name{process_one_gcm2}
\alias{process_one_gcm2}
\title{Process one GCM at a time}
\usage{
process_one_gcm2(
  gcm_nm,
  ssp,
  bbox,
  period,
  max_run,
  dbnames = dbnames,
  dbCon,
  cache
)
}
\arguments{
\item{gcm_nm}{character. Name of GCM to use.}

\item{ssp}{character. Vector of SSP-RCP scenarios (representative concentration pathways paired with shared socioeconomic pathways).
Options are \code{\link[=list_ssp]{list_ssp()}}. Defaults to all scenarios available.}

\item{bbox}{numeric. Vector of length 4 giving bounding box of study region, created by \code{get_bb()}.}

\item{period}{character. Vector of labels of the periods to use.
Can be obtained from \code{\link[=list_gcm_period]{list_gcm_period()}}. Defaults to all periods available.}

\item{max_run}{integer. Maximum number of model runs to include.
A value of 0 returns the \code{ensembleMean} only. Runs are included in the order they
are found in the models data until \code{max_run} is reached. Defaults to 0L.}

\item{dbnames}{\code{data.frame} with the list of available GCMs and their
corresponding names in the PostGIS data base.  See climr:::dbnames}

\item{dbCon}{A db connection object created by \code{data_connect}.}

\item{cache}{logical. Specifying whether to cache new data locally or no. Defaults to \code{TRUE}.}
}
\value{
\code{SpatRaster}
}
\description{
Process one GCM at a time
}
