% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{threaded_downscale_}
\alias{threaded_downscale_}
\title{Wrapper function for \code{downscale_} for parallelising}
\usage{
threaded_downscale_(
  xyz,
  normal,
  gcm,
  gcm_ts,
  gcm_hist,
  historic,
  historic_ts,
  ...
)
}
\arguments{
\item{xyz}{a \code{data.frame} with the following columns "long", "lat", "elev",
and a unique "id". Any extra columns will be ignored and not output.}

\item{normal}{\code{SpatRaster}. Reference normal baseline input from \code{normal_input}.}

\item{gcm}{\code{list} of \code{SpatRasters}. Global Circulation Models outputs from \code{\link[=gcm_input]{gcm_input()}}. Default to \code{NULL}.}

\item{gcm_ts}{\code{list} of \code{SpatRasters}. TODO}

\item{gcm_hist}{\code{list} of \code{SpatRasters}. TODO}

\item{historic}{\code{list} of \code{SpatRasters}. Historic time period outputs from \code{\link[=historic_input]{historic_input()}}. Default to \code{NULL}.}

\item{historic_ts}{\code{list} of \code{SpatRasters}. TODO}

\item{...}{further arguments passed to \code{downscale_}}
}
\value{
A \code{data.table}
}
\description{
Wrapper function for \code{downscale_} for parallelising
}
