% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{threaded_downscale_}
\alias{threaded_downscale_}
\title{TODO: fill documentation here}
\usage{
threaded_downscale_(normal_path, gcm_paths, historic_paths, ...)
}
\arguments{
\item{normal_path}{TODO}

\item{gcm_paths}{TODO}

\item{historic_paths}{TODO}

\item{...}{further arguments passed to \code{downscale_}}
}
\value{
A \code{data.table}
}
\description{
TODO: fill documentation here
}
