% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{threaded_downscale_}
\alias{threaded_downscale_}
\title{TODO: fill documentation here}
\usage{
threaded_downscale_(xyz, normal_path, gcm_paths, historic_paths, vars, ppt_lr)
}
\arguments{
\item{xyz}{}

\item{normal_path}{}

\item{gcm_paths}{}

\item{historic_paths}{}

\item{vars}{}

\item{ppt_lr}{}
}
\value{
\code{data.table}
}
\description{
TODO: fill documentation here
}
