% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{threaded_downscale_}
\alias{threaded_downscale_}
\title{Wrapper function for \code{downscale_} for parallelising}
\usage{
threaded_downscale_(
  xyz,
  normal,
  gcm,
  gcm_ts,
  gcm_hist,
  historic,
  historic_ts,
  ...
)
}
\arguments{
\item{xyz}{A 3-column matrix or data.frame (x, y, z) or (lon, lat, elev).}

\item{normal}{Reference normal baseline input from \code{normal_input}.}

\item{gcm}{Global Circulation Models input from \code{gcm_input}. Default to NULL.}

\item{gcm_ts}{TODO}

\item{gcm_hist}{TODO}

\item{historic}{Historic time period input from \code{historic_input}. Default to NULL}

\item{historic_ts}{TODO}

\item{...}{further arguments passed to \code{downscale_}}
}
\value{
A \code{data.table}
}
\description{
Wrapper function for \code{downscale_} for parallelising
}
