% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{unpackRasters}
\alias{unpackRasters}
\title{Pack rasters for parallel computing}
\usage{
unpackRasters(ras)
}
\arguments{
\item{ras}{a \code{SpatRaster} or list of \code{SpatRasters}, or \code{NULL}}
}
\value{
\code{NULL}, a \code{PackedSpatRaster} or list of \code{PackedSpatRasters}
}
\description{
Pack rasters for parallel computing
}
