% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{vancouver}
\alias{vancouver}
\title{Example line data}
\format{
\code{PackedSpatVector}
}
\source{
todo add/describe source
}
\usage{
vancouver
}
\description{
Shapefile of coast of Vancouver, BC, Canada. Mostly
used for plotting.
To use this data set, you need to \code{terra::unwrap}
it first, e.g.:
\code{vancouver <- terra::unwrap(get(data(vancouver)))}
}
\keyword{datasets}
